Data Structure

Data Structure

Variables and factors can potentially contain a lot of data, so DFG has different structures that are specialized for each use-case and level of detail. For example, if you want to retrieve just a simple summary of the structure, you can use the summary or skeleton structures to identify variables and factors of interest and then retrieve more detail on the subset.

Note that drivers support all of the structures.

The following section discusses the datastructures for each level. A quick summary of the types and the available properties (some of which are derived) is provided below.

Accessible properties for each of the variable structures:

| | Label | Timestamp | Tags | Estimates | Soft Type | Solvable | Solver Data | Small Data | Big Data Entries | |––––––––––-|–––-|–––––-|–––|–––––-|–––––-|–––––|––––––-|––––––|–––––––––| | SkeletonDFGVariable | X | | X | | | | | | | | DFGVariableSummary | X | X | X | X | Symbol | | | | X | | DFGVariable | X | X | X | X | X | X | X | X | X |

Accessible properties for each of the factor structures:

| | Label | Timestamp | Tags | Factor Type | Solvable | Solver Data | |–––––––––-|–––-|–––––-|–––|––––––-|–––––|––––––-| | SkeletonDFGFactor | X | | X | | | | | DFGFactorSummary | X | X | X | | | | | DFGFactor | X | X | X | X | X | X |

DFG Skeleton

DistributedFactorGraphs.SkeletonDFGVariable — Type.

struct SkeletonDFGVariable <: AbstractDFGVariable

Skeleton variable structure for a DistributedFactorGraph variable.

Fields:

label::Symbol

Variable label, e.g. :x1. Accessor: getLabel

tags::Set{Symbol}

Variable tags, e.g [:POSE, :VARIABLE, and :LANDMARK]. Accessors: getTags, mergeTags!, and removeTags!

source

DistributedFactorGraphs.SkeletonDFGFactor — Type.

struct SkeletonDFGFactor <: AbstractDFGFactor

Skeleton factor structure for a DistributedFactorGraph factor.

Fields:

label::Symbol

Factor label, e.g. :x1f1. Accessor: getLabel

tags::Set{Symbol}

Factor tags, e.g [:FACTOR]. Accessors: getTags, mergeTags!, and removeTags!

_variableOrderSymbols::Array{Symbol,1}

Internal cache of the ordering of the neighbor variables. Rather use getNeighbors to get the list as this is an internal value. Accessors: getVariableOrder

source

DFG Summary

DistributedFactorGraphs.DFGVariableSummary — Type.

struct DFGVariableSummary <: AbstractDFGVariable

Summary variable structure for a DistributedFactorGraph variable.

Fields:

label::Symbol

Variable label, e.g. :x1. Accessor: getLabel

timestamp::Dates.DateTime

Variable timestamp. Accessors: getTimestamp, setTimestamp

tags::Set{Symbol}

Variable tags, e.g [:POSE, :VARIABLE, and :LANDMARK]. Accessors: getTags, mergeTags!, and removeTags!

ppeDict::Dict{Symbol,#s25} where #s25<:AbstractPointParametricEst

Dictionary of parametric point estimates keyed by solverDataDict keys Accessors: addPPE!, updatePPE!, and deletePPE!

softtypename::Symbol

Symbol for the softtype for the underlying variable. Accessor: getSofttype

bigData::Dict{Symbol,AbstractBigDataEntry}

Dictionary of large data associated with this variable. Accessors: addBigDataEntry!, getBigDataEntry, updateBigDataEntry!, and deleteBigDataEntry!

source

DistributedFactorGraphs.DFGFactorSummary — Type.

struct DFGFactorSummary <: AbstractDFGFactor

Read-only summary factor structure for a DistributedFactorGraph factor.

Fields:

label::Symbol

Factor label, e.g. :x1f1. Accessor: getLabel

timestamp::Dates.DateTime

Variable timestamp. Accessors: getTimestamp

tags::Set{Symbol}

Factor tags, e.g [:FACTOR]. Accessors: getTags, mergeTags!, and removeTags!

_variableOrderSymbols::Array{Symbol,1}

Internal cache of the ordering of the neighbor variables. Rather use getNeighbors to get the list as this is an internal value. Accessors: getVariableOrder

source

Full DFG Node

DistributedFactorGraphs.DFGVariable — Type.

struct DFGVariable{T<:InferenceVariable} <: AbstractDFGVariable

Complete variable structure for a DistributedFactorGraph variable.

Fields:

label::Symbol

Variable label, e.g. :x1. Accessor: getLabel

timestamp::Dates.DateTime

Variable timestamp. Accessors: getTimestamp, setTimestamp

nstime::Dates.Nanosecond

Nano second time, for more resolution on timestamp (only subsecond information)

tags::Set{Symbol}

Variable tags, e.g [:POSE, :VARIABLE, and :LANDMARK]. Accessors: getTags, mergeTags!, and removeTags!

ppeDict::Dict{Symbol,#s18} where #s18<:AbstractPointParametricEst

Dictionary of parametric point estimates keyed by solverDataDict keys Accessors: addPPE!, updatePPE!, and deletePPE!

solverDataDict::Dict{Symbol,VariableNodeData{T}} where T<:InferenceVariable

Dictionary of solver data. May be a subset of all solutions if a solver key was specified in the get call. Accessors: addVariableSolverData!, updateVariableSolverData!, and deleteVariableSolverData!

smallData::Dict{String,String}

Dictionary of small data associated with this variable. Accessors: getSmallData, setSmallData!

bigData::Dict{Symbol,AbstractBigDataEntry}

Dictionary of large data associated with this variable. Accessors: addBigDataEntry!, getBigDataEntry, updateBigDataEntry!, and deleteBigDataEntry!

solvable::Base.RefValue{Int64}

Solvable flag for the variable. Accessors: getSolvable, setSolvable!

source

DistributedFactorGraphs.DFGFactor — Type.

struct DFGFactor{T, N} <: AbstractDFGFactor

Complete factor structure for a DistributedFactorGraph factor.

Fields:

label::Symbol

Factor label, e.g. :x1f1. Accessor: getLabel

timestamp::Dates.DateTime

Variable timestamp. Accessors: getTimestamp, setTimestamp

nstime::Dates.Nanosecond

Nano second time, for more resolution on timestamp (only subsecond information)

tags::Set{Symbol}

Factor tags, e.g [:FACTOR]. Accessors: getTags, mergeTags!, and removeTags!

solverData::Base.RefValue{GenericFunctionNodeData{T}} where T

Solver data. Accessors: getSolverData, setSolverData!

solvable::Base.RefValue{Int64}

Solvable flag for the factor. Accessors: getSolvable, setSolvable!

_variableOrderSymbols::Tuple{Vararg{Symbol,N}} where N

Internal cache of the ordering of the neighbor variables. Rather use getVariableOrder to get the list as this is an internal value. Accessors: getVariableOrder

source

Additional Offloaded Data

Additional, larger data can be associated with variables using keyed big data entries.