Using Graph Data

Using Graph Elements

Variables and factors in DistributedFactorGraphs are used for a variety of different applications. We have tried to compartmentalize the data as much as possible so that users do not need to dig around to find what they need (it's a work in progress).

There are three fundamental types of data in DFG:

Variable and factor data (stored in the nodes themselves)

Offloaded big data elements (keyed in a variable or factor, but stored in another location)

Graph data (data that is related to the graph itself)

The following is a guideline to using these parameters.

NOTE: Some functions are direct accessors to the internal parameters, others are derived functions (e.g. getLabel(v) = v.label). In other cases the accessors are simplified ways to interact with the structures. We recommend using the accessors as the internal structure may change over time.

NOTE: Adds in general throw an error if the element already exists. Update will update the element if it exists, otherwise it will add it.

NOTE: In general these functions will return an error if the respective element is not found. This is to avoid returning, say, nothing, which will be horribly confusing if you tried getVariableSolverData(dfg, :a, :b) and it returned nothing - which was missing, :a or :b, or was there a communication issue? We recommend coding defensively and trapping errors in critical portions of your user code.

NOTE: All data is passed by reference, so if you update the returned structure it will update in the graph. The database driver is an exception, and once the variable or factor is updated you need to call update* to persist the changes to the graph.

The following examples make use this data:

using IncrementalInference
# Create a DFG with default solver parameters using the LightGraphs.jl driver.
dfg = LightDFG{SolverParams}(params=SolverParams())

x0 = addVariable!(dfg, :x0, ContinuousScalar, labels = [:POSE], solvable=1)
x1 = addVariable!(dfg, :x1, ContinuousScalar, labels = [:POSE], solvable=1)
f1 = addFactor!(dfg, [:x0; :x1], LinearConditional(Normal(50.0,2.0)), solvable=1)

Variable and Factor Elements

Common Elements

Labels

Labels are the principle identifier of a variable or factor.

DistributedFactorGraphs.getLabel — Function.

getLabel(v)

Return the label for a DFGNode.

source

Timestamps

Each variable or factor can have a timestamp associated with it.

DistributedFactorGraphs.getTimestamp — Function.

getTimestamp(v)

Get the timestamp of a DFGNode.

source

DistributedFactorGraphs.setTimestamp — Function.

setTimestamp(v, ts; verbose)

Set the timestamp of a DFGVariable object returning a new DFGVariable. Note: Since the timestamp field is not mutable setTimestamp returns a new variable with the updated timestamp (note the absence of !). Use updateVariable! on the returened variable to update it in the factor graph if needed. Alternatively use setTimestamp!. See issue #315.

source

Tags

Tags are a set of symbols that contain identifiers for the variable or factor.

DistributedFactorGraphs.listTags — Function.

listTags(dfg, sym)

Return the tags for a variable or factor.

source

DistributedFactorGraphs.mergeTags! — Function.

mergeTags!(dfg, sym, tags)

Merge add tags to a variable or factor (union)

source

DistributedFactorGraphs.removeTags! — Function.

removeTags!(dfg, sym, tags)

Remove the tags from the node (setdiff)

source

DistributedFactorGraphs.emptyTags! — Function.

emptyTags!(dfg, sym)

Empty all tags from the node (empty)

source

Solvable

The solvable flag indicates whether the solver should make use of the variable or factor while solving the graph. This can be used to construct graphs in chunks while solving asynchronously, or for selectively solving portions of the graph.

DistributedFactorGraphs.getSolvable — Function.

getSolvable(var)

Variables or factors may or may not be 'solvable', depending on a user definition. Useful for ensuring atomic transactions.

Related:

isSolveInProgress

source

getSolvable(dfg, sym)

Get 'solvable' parameter for either a variable or factor.

source

DistributedFactorGraphs.setSolvable! — Function.

setSolvable!(node, solvable)

Set the solvable parameter for either a variable or factor.

source

setSolvable!(dfg, sym, solvable)

Set the solvable parameter for either a variable or factor.

source

Variables

Soft Type

The soft type is the underlying inference variable type, such as a Pose2.

DistributedFactorGraphs.getSofttype — Function.

getSofttype(vnd)

Variable nodes softtype information holding a variety of meta data associated with the type of variable stored in that node of the factor graph.

Related

getVariableType

source

Packed Parametric Estimates

Solved graphs contain packed parametric estimates for the variables, which are keyed by the solution (the default is saved as :default).

For each PPE structure, there are accessors for getting individual values:

DistributedFactorGraphs.getMaxPPE — Function.

getMaxPPE(est)

source

DistributedFactorGraphs.getMeanPPE — Function.

getMeanPPE(est)

source

DistributedFactorGraphs.getSuggestedPPE — Function.

getSuggestedPPE(est)

source

Related functions for getting, adding/updating, and deleting PPE structures:

Missing docstring.

Missing docstring for listPPE. Check Documenter's build log for details.

DistributedFactorGraphs.getPPE — Function.

getPPE(vari)
getPPE(vari, solveKey)

Get the parametric point estimate (PPE) for a variable in the factor graph.

Notes

Defaults on keywords solveKey and method

Related

getMeanPPE, getMaxPPE, getKDEMean, getKDEFit, getPPEs, getVariablePPEs

source

getPPE(dfg, variablekey)
getPPE(dfg, variablekey, ppekey)

Get the parametric point estimate (PPE) for a variable in the factor graph for a given solve key.

Notes

Defaults on keywords solveKey and method

Related getMeanPPE, getMaxPPE, getKDEMean, getKDEFit, getPPEs, getVariablePPEs

source

DistributedFactorGraphs.addPPE! — Function.

addPPE!(dfg, variablekey, ppe)
addPPE!(dfg, variablekey, ppe, ppekey)

Add variable PPE, errors if it already exists.

source

addPPE!(dfg, sourceVariable)
addPPE!(dfg, sourceVariable, ppekey)

Add a new PPE entry from a deepcopy of the source variable PPE. NOTE: Copies the solver data.

source

DistributedFactorGraphs.updatePPE! — Function.

Update PPE data if it exists, otherwise add it – one call per key::Symbol=:default.

source

updatePPE!(dfg, sourceVariable)
updatePPE!(dfg, sourceVariable, ppekey)

Update PPE data if it exists, otherwise add it. NOTE: Copies the PPE data.

source

updatePPE!(dfg, sourceVariables)
updatePPE!(dfg, sourceVariables, ppekey)

Update PPE data if it exists, otherwise add it.

source

DistributedFactorGraphs.deletePPE! — Function.

deletePPE!(dfg, variablekey)
deletePPE!(dfg, variablekey, ppekey)

Delete PPE data, returns the deleted element.

source

deletePPE!(dfg, sourceVariable)
deletePPE!(dfg, sourceVariable, ppekey)

Delete PPE data, returns the deleted element.

source

DistributedFactorGraphs.mergePPEs! — Function.

mergePPEs!(destVariable, sourceVariable)

Merges and updates solver and estimate data for a variable (variable can be from another graph). Note: Makes a copy of the estimates and solver data so that there is no coupling between graphs.

source

Example of PPE operations:

# Add a new PPE of type MeanMaxPPE to :x0
ppe = MeanMaxPPE(:default, [0.0], [0.0], [0.0])
addPPE!(dfg, :x0, ppe)
@show listPPEs(dfg, :x0)
# Get the data back - note that this is a reference to above.
v = getPPE(dfg, :x0, :default)
# Delete it
deletePPE!(dfg, :x0, :default)
# Update add it
updatePPE!(dfg, :x0, ppe, :default)
# Update update it
updatePPE!(dfg, :x0, ppe, :default)
# Bulk copy PPE's for x0 and x1
updatePPE!(dfg, [x0], :default)

Solver Data

Solver data is used by IncrementalInference/RoME/Caesar solver to produce the above PPEs.

Related functions:

DistributedFactorGraphs.listVariableSolverData — Function.

listVariableSolverData(dfg, variablekey)

List all the solver data keys in the variable.

source

DistributedFactorGraphs.getVariableSolverData — Function.

getVariableSolverData(dfg, variablekey)
getVariableSolverData(dfg, variablekey, solvekey)

Get variable solverdata for a given solve key.

source

DistributedFactorGraphs.addVariableSolverData! — Function.

addVariableSolverData!(dfg, variablekey, vnd)
addVariableSolverData!(dfg, variablekey, vnd, solvekey)

Add variable solver data, errors if it already exists.

source

addVariableSolverData!(dfg, sourceVariable)
addVariableSolverData!(dfg, sourceVariable, solvekey)

Add a new solver data entry from a deepcopy of the source variable solver data. NOTE: Copies the solver data.

source

DistributedFactorGraphs.updateVariableSolverData! — Function.

updateVariableSolverData!(dfg, variablekey, vnd)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey, useCopy)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey, useCopy, fields)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey, useCopy, fields, verbose)

Update variable solver data if it exists, otherwise add it.

Notes:

useCopy=true to copy solver data and keep separate memory.

Use fields to updated only a few VND.fields while adhering to useCopy.

Related

mergeVariableSolverData!

source

DistributedFactorGraphs.deleteVariableSolverData! — Function.

deleteVariableSolverData!(dfg, variablekey)
deleteVariableSolverData!(dfg, variablekey, solvekey)

Delete variable solver data, returns the deleted element.

source

deleteVariableSolverData!(dfg, sourceVariable)
deleteVariableSolverData!(dfg, sourceVariable, solvekey)

Delete variable solver data, returns the deleted element.

source

DistributedFactorGraphs.mergeVariableSolverData! — Function.

mergeVariableSolverData!(destVariable, sourceVariable)

Merges and updates solver and estimate data for a variable (variable can be from another graph). If the same key is present in another collection, the value for that key will be the value it has in the last collection listed (updated). Note: Makes a copy of the estimates and solver data so that there is no coupling between graphs.

source

Example of solver data operations:

# Add new VND of type ContinuousScalar to :x0
# Could also do VariableNodeData(ContinuousScalar())
vnd = VariableNodeData{ContinuousScalar}()
addVariableSolverData!(dfg, :x0, vnd, :parametric)
@show listVariableSolverData(dfg, :x0)
# Get the data back - note that this is a reference to above.
vndBack = getVariableSolverData(dfg, :x0, :parametric)
# Delete it
deleteVariableSolverData!(dfg, :x0, :parametric)

Small Data

Small data allows you to assign a dictionary to variables. It is a useful way to keep small amounts of string data in a variable. As it is stored in the graph itself, large entries will slow the graph down, so if data should exceed a few bytes/kb, it should rather be saved in bigData.

DistributedFactorGraphs.getSmallData — Function.

getSmallData(v)

Get the small data for a variable.

source

DistributedFactorGraphs.setSmallData! — Function.

setSmallData!(v, smallData)

Set the small data for a variable. This will overwrite old smallData.

source

Example:

setSmallData!(x0, Dict("entry"=>"entry value"))
getSmallData(x0)

Big Data

Factors

Graph-Related Data

DFG can store data in the graph itself (as opposed to inside graph elements). When you retrieve graphs from a database, this information is carried along. If you are working with an in-memory graph, the structure is flattened into the graph itself as userData, robotData, and sessionData.

Graphs reside inside a hierarchy made up in the following way:

User1

Robot1

Session1 (the graph itself)

User2

Robot2

Robot3

Session2

Session3

This data can be retrieved with the follow functions:

DistributedFactorGraphs.getUserData — Function.

getUserData(dfg)

Get the user data associated with the graph.

source

DistributedFactorGraphs.getRobotData — Function.

getRobotData(dfg)

Get the robot data associated with the graph.

source

DistributedFactorGraphs.getSessionData — Function.

getSessionData(dfg)

Get the session data associated with the graph.

source

It can be set using the following functions:

DistributedFactorGraphs.setUserData! — Function.

setUserData!(dfg, data)

Set the user data associated with the graph.

source

DistributedFactorGraphs.setRobotData! — Function.

setRobotData!(dfg, data)

Set the robot data associated with the graph.

source

DistributedFactorGraphs.setSessionData! — Function.

setSessionData!(dfg, data)

Set the session data associated with the graph.

source

Example of using graph-level data:

setUserData!(dfg, Dict(:a => "Hello"))
getUserData(dfg)