Traversing and Querying

Traversing and Querying a Graph

DistributedFactorGraphs.getNeighbors — Function.

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.getSubgraphAroundNode — Function.

getSubgraphAroundNode(dfg, node)
getSubgraphAroundNode(dfg, node, distance)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors, addToDFG; solvable)

Retrieve a deep subgraph copy around a given variable or factor. Optionally provide a distance to specify the number of edges should be followed. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables. Note: Always returns the node at the center, but filters around it if solvable is set.

source

DistributedFactorGraphs.getSubgraph — Function.

getSubgraph(dfg, variableFactorLabels)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors, addToDFG)

Get a deep subgraph copy from the DFG given a list of variables and factors. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

Missing docstring.

Missing docstring for getIncidenceMatrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getIncidenceMatrixSparse. Check Documenter's build log for details.