Reference

Function Reference

Function Reference

Graph Types

Creating DFG Factor Graphs

Getting Factor Graph Nodes

Updating Factor Graph Nodes

Deleting Factor Graph Nodes

Getting Adjacency Matrix

Validating Factor Graph Connectivity

Copying Subgraphs

Visualization

DataFrame Extension Functions

Graph Types

GraphsDFG

Creating DFG Factor Graphs

DistributedFactorGraphs.addVariable! — Function.

addVariable!(dfg, variable)

Add a DFGVariable to a DFG.

source

addVariable!(dfg, variable)

Add a DFGVariable to a DFG.

source

DistributedFactorGraphs.addFactor! — Function.

addFactor!(dfg, variables, factor)

Add a DFGFactor to a DFG.

source

addFactor!(dfg, variableIds, factor)

Add a DFGFactor to a DFG.

source

addFactor!(dfg, variables, factor)

Add a DFGFactor to a DFG.

source

addFactor!(dfg, variableIds, factor)

Add a DFGFactor to a DFG.

source

Getting Factor Graph Nodes

DistributedFactorGraphs.getVariables — Function.

getVariables(dfg)
getVariables(dfg, regexFilter; tags)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables.

source

getVariables(dfg)
getVariables(dfg, regexFilter; tags)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables.

source

DistributedFactorGraphs.getFactors — Function.

getFactors(dfg)
getFactors(dfg, regexFilter)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

getFactors(dfg)
getFactors(dfg, regexFilter)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

DistributedFactorGraphs.getVariable — Function.

getVariable(dfg, variableId)

Get a DFGVariable from a DFG using its underlying integer ID.

source

getVariable(dfg, label)

Get a DFGVariable from a DFG using its label.

source

getVariable(dfg, variableId)

Get a DFGVariable from a DFG using its underlying integer ID.

source

getVariable(dfg, label)

Get a DFGVariable from a DFG using its label.

source

DistributedFactorGraphs.getFactor — Function.

getFactor(dfg, factorId)

Get a DFGFactor from a DFG using its underlying integer ID.

source

getFactor(dfg, label)

Get a DFGFactor from a DFG using its label.

source

getFactor(dfg, factorId)

Get a DFGFactor from a DFG using its underlying integer ID.

source

getFactor(dfg, label)

Get a DFGFactor from a DFG using its label.

source

DistributedFactorGraphs.getNeighbors — Function.

getNeighbors(dfg, node; ready, backendset)

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

getNeighbors(dfg, label; ready, backendset)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

getNeighbors(dfg, node; ready, backendset)

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

getNeighbors(dfg, label; ready, backendset)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.ls — Function.

ls(dfg)
ls(dfg, regexFilter; tags)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables.

source

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

ls(dfg, label)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

ls(dfg)
ls(dfg, regexFilter; tags)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables.

source

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

ls(dfg, label)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.lsf — Function.

lsf(dfg, label)

Alias for getNeighbors - returns neighbors around a given node label.

source

lsf(dfg, label)

Alias for getNeighbors - returns neighbors around a given node label.

source

Updating Factor Graph Nodes

DistributedFactorGraphs.updateVariable! — Function.

updateVariable!(dfg, variable)

Update a complete DFGVariable in the DFG.

source

updateVariable!(dfg, variable)

Update a complete DFGVariable in the DFG.

source

DistributedFactorGraphs.updateFactor! — Function.

updateFactor!(dfg, factor)

Update a complete DFGFactor in the DFG.

source

updateFactor!(dfg, factor)

Update a complete DFGFactor in the DFG.

source

Deleting Factor Graph Nodes

DistributedFactorGraphs.deleteVariable! — Function.

deleteVariable!(dfg, label)

Delete a DFGVariable from the DFG using its label.

source

deleteVariable!(dfg, variable)

Delete a referenced DFGVariable from the DFG.

source

deleteVariable!(dfg, label)

Delete a DFGVariable from the DFG using its label.

source

deleteVariable!(dfg, variable)

Delete a referenced DFGVariable from the DFG.

source

DistributedFactorGraphs.deleteFactor! — Function.

deleteFactor!(dfg, label)

Delete a DFGFactor from the DFG using its label.

source

deleteFactor!(dfg, factor)

Delete the referened DFGFactor from the DFG.

source

deleteFactor!(dfg, label)

Delete a DFGFactor from the DFG using its label.

source

deleteFactor!(dfg, factor)

Delete the referened DFGFactor from the DFG.

source

Getting Adjacency Matrix

DistributedFactorGraphs.getAdjacencyMatrix — Function.

getAdjacencyMatrix(dfg)

Get an adjacency matrix for the DFG, returned as a Matrix{Union{Nothing, Symbol}}. Rows are all factors, columns are all variables, and each cell contains either nothing or the symbol of the relating factor. The first row and first column are factor and variable headings respectively.

source

getAdjacencyMatrix(dfg)

Get an adjacency matrix for the DFG, returned as a Matrix{Union{Nothing, Symbol}}. Rows are all factors, columns are all variables, and each cell contains either nothing or the symbol of the relating factor. The first row and first column are factor and variable headings respectively.

source

Validating Factor Graph Connectivity

DistributedFactorGraphs.isFullyConnected — Function.

isFullyConnected(dfg)

Checks if the graph is fully connected, returns true if so.

source

isFullyConnected(dfg)

Checks if the graph is fully connected, returns true if so.

source

DistributedFactorGraphs.hasOrphans — Function.

hasOrphans(dfg)

Checks if the graph is not fully connected, returns true if it is not contiguous.

source

hasOrphans(dfg)

Checks if the graph is not fully connected, returns true if it is not contiguous.

source

Copying Subgraphs

DistributedFactorGraphs.getSubgraphAroundNode — Function.

getSubgraphAroundNode(dfg, node)
getSubgraphAroundNode(dfg, node, distance)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors, addToDFG)

Retrieve a deep subgraph copy around a given variable or factor. Optionally provide a distance to specify the number of edges should be followed. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors, addToDFG)
getSubgraphAroundNode(dfg, node)
getSubgraphAroundNode(dfg, node, distance)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors)

Retrieve a deep subgraph copy around a given variable or factor. Optionally provide a distance to specify the number of edges should be followed. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

DistributedFactorGraphs.getSubgraph — Function.

getSubgraph(dfg, variableFactorLabels)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors, addToDFG)

Get a deep subgraph copy from the DFG given a list of variables and factors. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

getSubgraph(dfg, variableFactorLabels)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors, addToDFG)

Get a deep subgraph copy from the DFG given a list of variables and factors. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

Visualization

DistributedFactorGraphs.toDot — Function.

toDot(dfg)

Produces a dot-format of the graph for visualization.

source

toDot(dfg)

Produces a dot-format of the graph for visualization.

source

DistributedFactorGraphs.toDotFile — Function.

toDotFile(dfg)
toDotFile(dfg, fileName)

Produces a dot file of the graph for visualization. Download XDot to see the data

Note

Default location "/tmp/dfg.dot" – MIGHT BE REMOVED

Can be viewed with the xdot system application.

Based on graphviz.org

source

toDotFile(dfg)
toDotFile(dfg, fileName)

Produces a dot file of the graph for visualization. Download XDot to see the data

Note

Default location "/tmp/dfg.dot" – MIGHT BE REMOVED

Can be viewed with the xdot system application.

Based on graphviz.org

source

DataFrame Extension Functions

getAdjacencyMatrixDataFrame