Traversing and Querying

Traversing and Querying a Graph

DistributedFactorGraphs.getNeighbors — Function.

getNeighbors(dfg, label; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

Missing docstring.

Missing docstring for getSubgraphAroundNode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getSubgraph. Check Documenter's build log for details.

DistributedFactorGraphs.getBiadjacencyMatrix — Function.

getBiadjacencyMatrix(dfg; solvable)

Get a matrix indicating adjacency between variables and factors. Returned as a named tuple: B::SparseMatrixCSC{Int}, varLabels::Vector{Symbol) facLabels::Vector{Symbol). Rows are the factors, columns are the variables, with the corresponding labels in varLabels,facLabels.

source