Blobs and Blobentries

Blobs and Blobentries

Types

Functions

Types

DistributedFactorGraphs.Blobentry — Type

struct Blobentry

A Blobentry is a small about of structured data that holds reference information to find an actual blob. Many Blobentrys can exist on different graph nodes spanning Agents and Factor Graphs which can all reference the same Blob.

Notes:

blobids should be unique within a blobstore and are immutable.

DistributedFactorGraphs.AbstractBlobstore — Type

AbstractBlobstore{T}

Abstract supertype for all blobstore implementations.

Usage

Subtypes of AbstractBlobstore{T} must implement the required interface for blob storage and retrieval, such as:

add!(store, blobid, blob): Add a new blob to the store.

get(store, blobid): Retrieve a blob by its ID.

list(store): List all blob IDs in the store.

The parameter T represents the type of blobs stored (e.g., Vector{UInt8} or a custom Blob type).

See concrete implementations for details.

Design Notes

blobid is not considered unique across blobstores with different labels only within a single blobstore.

We cannot guarantee that blobid is unique across different blobstores with the same label and this is up to the end user.

Within a single blobstore addBlob! will fail if there is a UUID collision.

TODO: We should consider using uuid7 for blobids (requires jl v1.12).

Blobstroresare identified by a label::Symbol, which allows for multiple blobstores to coexist in the same system.

TODO: If we want to make the blobid=>Blob pair immutable:

We can use the tombstone pattern to mark a blob as deleted. See FolderStore in PR#TODO.

Design goal: all Blobstores with the same label can contain the same blobid=>Blob pair and the blobs should be identical since they are immutable.

Functions

DistributedFactorGraphs.addBlobentry! — Method

addBlobentry!(node, entry)

DistributedFactorGraphs.checkHash — Method

checkHash(entry::Blobentry, blob) -> Union{Bool,Nothing}

Checks the integrity of a blob against the hashes (crc32c, sha256) stored in the given Blobentry.

Returns true if all present hashes (crchash, shahash) match the computed values from blob.

Returns false if any present hash does not match.

Returns nothing if no hashes are stored in the Blobentry to check against.

DistributedFactorGraphs.deleteBlobentry! — Method

deleteBlobentry!(node, label)

DistributedFactorGraphs.getBlobentry — Method

getBlobentry(node, label)

DistributedFactorGraphs.getfirstBlobentry — Method

getfirstBlobentry(
    node;
    labelFilter,
    blobidFilter,
    sortby,
    sortlt
)

Finds and returns the first blob entry that matches the filter. The result is sorted by sortby[=getLabel] and sortlt[=natural_lt] before returning the first entry. Also see: getBlobentry

DistributedFactorGraphs.hasBlobentry — Method

hasBlobentry(node, label)

Does a blob entry exist with label.

DistributedFactorGraphs.incrDataLabelSuffix — Method

incrDataLabelSuffix(dfg, vla, bllb; datalabel)

If the blob label datalabel already exists, then this function will return the name datalabel_1. If the blob label datalabel_1 already exists, then this function will return the name datalabel_2.

DistributedFactorGraphs.listBlobentries — Method

listBlobentries(node)

List all Blobentry keys for a variable label in dfg

DistributedFactorGraphs.listBlobentrySequence — Function

listBlobentrySequence(dfg, lb, pattern)
listBlobentrySequence(dfg, lb, pattern, _sort)

List a collection of blob entries per variable that match a particular pattern::Regex.

Notes

Optional sort function argument, default is unsorted.

Likely use of sortDFG for basic Symbol sorting.

Example

listBlobentrySequence(fg, :x0, r"IMG_CENTER", sortDFG)
15-element Vector{Symbol}:
 :IMG_CENTER_21676
 :IMG_CENTER_21677
 :IMG_CENTER_21678
 :IMG_CENTER_21679
...

DistributedFactorGraphs.mergeBlobentry! — Method

mergeBlobentry!(node, entry)

DistributedFactorGraphs.addBlob! — Function

Adds a blob to the blob store or dfg with the blobid.

Related addBlobentry! Implement addBlob!(store::AbstractBlobstore, blobid::UUID, data)

addBlob!(dfg, entry, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:68.

addBlob!(store, entry, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:72.

addBlob!(store, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:77.

addBlob!(store, blobid, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:154.

addBlob!(store, blobid, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:226.

addBlob!(store, blobid, linkfile)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:277.

addBlob!(store, blobid, blob)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:361.

DistributedFactorGraphs.deleteBlob! — Function

Delete a blob from the blob store or dfg with the given entry.

Related deleteBlobentry! Implement deleteBlob!(store::AbstractBlobstore, blobid::UUID)

deleteBlob!(dfg, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:80.

deleteBlob!(store, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:84.

deleteBlob!(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:166.

deleteBlob!(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:234.

deleteBlob!(store)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:288.

deleteBlob!(store, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:292.

deleteBlob!(store, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:293.

deleteBlob!(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:369.

DistributedFactorGraphs.getBlob — Function

Get the data blob for the specified blobstore or dfg.

Related getBlobentry Implement getBlob(store::AbstractBlobstore, blobid::UUID)

getBlob(dfg, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:57.

getBlob(store, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:63.

getBlob(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:140.

getBlob(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:219.

getBlob(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:269.

getBlob(store, blobid)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:354.

DistributedFactorGraphs.hasBlob — Function

Check if the blob store has a blob with the given blobid.

DistributedFactorGraphs.listBlobs — Function

List all blobids in the blob store. Implement listBlobs(store::AbstractBlobstore)

DistributedFactorGraphs.packBlob — Function

packBlob

Convert a file (JSON, JPG, PNG, BSON, LAS) to Vector{UInt8} for use as a Blob. Returns the blob and MIME type.

DistributedFactorGraphs.unpackBlob — Function

unpackBlob

Convert a Blob back to the origanal typ using the MIME type or DataFormat type.

DistributedFactorGraphs.deleteBlob_Agent! — Function

Convenience wrapper to delete a Blob from a Blobstore and its Blobentry from an agent.

deleteBlob_Agent!(dfg, entry_label)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:162.

DistributedFactorGraphs.deleteBlob_Graph! — Function

Convenience wrapper to delete a Blob from a Blobstore and its Blobentry from a graph.

deleteBlob_Graph!(dfg, entry_label)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:132.

DistributedFactorGraphs.deleteBlob_Variable! — Function

Convenience wrapper to delete a Blob form a Blobstore and its Blobentry from a variable.

deleteBlob_Variable!(dfg, variable_label, entry_label)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:102.

DistributedFactorGraphs.loadBlob_Agent — Function

Convenience wrapper to load a Blob for a given agent and Blobentry label.

loadBlob_Agent(dfg, entry_label)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:139.

DistributedFactorGraphs.loadBlob_Graph — Function

Convenience wrapper to load a Blob for a given graph and Blobentry label.

loadBlob_Graph(dfg, entry_label)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:109.

DistributedFactorGraphs.loadBlob_Variable — Function

Convenience wrapper to load a Blob for a given variable and Blobentry label.

loadBlob_Variable(dfg, variable_label, entry_label)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:66.

DistributedFactorGraphs.saveBlob_Agent! — Function

Convenience wrapper to save a Blob to a Blobstore and a Blobentry to an agent.

saveBlob_Agent!(dfg, blob, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:145.

saveBlob_Agent!(dfg, blob, entry_label; ...)
saveBlob_Agent!(
    dfg,
    blob,
    entry_label,
    blobstore;
    blobentry_kwargs...
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:151.

DistributedFactorGraphs.saveBlob_Graph! — Function

Convenience wrapper to save a Blob to a Blobstore and a Blobentry to a graph.

saveBlob_Graph!(dfg, blob, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:115.

saveBlob_Graph!(dfg, blob, entry_label; ...)
saveBlob_Graph!(
    dfg,
    blob,
    entry_label,
    blobstore;
    blobentry_kwargs...
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:121.

DistributedFactorGraphs.saveBlob_Variable! — Function

Convenience wrapper to save a Blob to a Blobstore and a Blobentry to a variable.

saveBlob_Variable!(dfg, variable_label, blob, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:79.

saveBlob_Variable!(
    dfg,
    variable_label,
    blob,
    entry_label;
    ...
)
saveBlob_Variable!(
    dfg,
    variable_label,
    blob,
    entry_label,
    blobstore;
    blobentry_kwargs...
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobWrappers.jl:90.