Blobs and BlobEntries

Blobs and BlobEntries

Types

Functions

Types

DistributedFactorGraphs.BlobEntry — Type

struct BlobEntry

A BlobEntry is a small about of structured data that holds reference information to find an actual blob. Many BlobEntrys can exist on different graph nodes spanning Agents and Factor Graphs which can all reference the same Blob.

Notes:

blobIds should be unique within a blobstore and are immutable.

Functions

DistributedFactorGraphs.addBlobEntry! — Method

addBlobEntry!(var, entry)

Add Data Entry to a DFG variable Should be extended if DFG variable is not returned by reference.

Also see: getBlobEntry, addBlob!, mergeBlobEntries!

DistributedFactorGraphs.buildSourceString — Method

buildSourceString(dfg, label)

Function to generate source string - agentLabel|graphLabel|varLabel

DistributedFactorGraphs.deleteBlobEntry! — Method

deleteBlobEntry!(var, key)

Delete a blob entry from the factor graph. Note this doesn't remove it from any data stores.

Notes:

users responsibility to delete data in db before deleting entry

DistributedFactorGraphs.getBlobEntries — Method

getBlobEntries(var)

Get blob entries, Vector{BlobEntry}

DistributedFactorGraphs.getBlobEntriesVariables — Method

getBlobEntriesVariables(
    dfg,
    bLblPattern;
    varList,
    dropEmpties
)

Get all blob entries matching a Regex pattern over variables

Notes

Use dropEmpties=true to not include empty lists in result.

Use keyword varList for which variables to search through.

DistributedFactorGraphs.getBlobEntry — Method

getBlobEntry(var, key)

Get data entry

Also see: addBlobEntry!, getBlob, listBlobEntries

DistributedFactorGraphs.getBlobEntryFirst — Method

getBlobEntryFirst(var, key)

Finds and returns the first blob entry that matches the regex.

Also see: getBlobEntry

DistributedFactorGraphs.hasBlobEntry — Method

hasBlobEntry(var, blobLabel)

Does a blob entry (element) exist with blobLabel.

DistributedFactorGraphs.incrDataLabelSuffix — Method

incrDataLabelSuffix(dfg, vla, bllb; datalabel)

If the blob label datalabel already exists, then this function will return the name datalabel_1. If the blob label datalabel_1 already exists, then this function will return the name datalabel_2.

DistributedFactorGraphs.listBlobEntries — Method

listBlobEntries(var)

List the blob entries associated with a particular variable.

DistributedFactorGraphs.listBlobEntrySequence — Function

listBlobEntrySequence(dfg, lb, pattern)
listBlobEntrySequence(dfg, lb, pattern, _sort)

List a collection of blob entries per variable that match a particular pattern::Regex.

Notes

Optional sort function argument, default is unsorted.

Likely use of sortDFG for basic Symbol sorting.

Example

listBlobEntrySequence(fg, :x0, r"IMG_CENTER", sortDFG)
15-element Vector{Symbol}:
 :IMG_CENTER_21676
 :IMG_CENTER_21677
 :IMG_CENTER_21678
 :IMG_CENTER_21679
...

DistributedFactorGraphs.mergeBlobEntries! — Method

mergeBlobEntries!(dst, dlbl, src, slbl, bllb)

Add a blob entry into the destination variable which already exists in a source variable.

See also: addBlobEntry!, getBlobEntry, listBlobEntries, getBlob

DistributedFactorGraphs.updateBlobEntry! — Method

updateBlobEntry!(var, bde)

Update data entry

DevNote

DF, unclear if update verb is applicable in this case, see #404

DistributedFactorGraphs.addBlob! — Function

Adds a blob to the blob store or dfg with the given entry.

Related addBlobEntry!

addBlob!(store, _, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:61.

addBlob!(dfg, entry, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:111.

addBlob!(store, entry, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:115.

addBlob!(store, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:121.

addBlob!(store, blobId, data, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:124.

addBlob!(store, data, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:128.

addBlob!(store, blobId, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:206.

addBlob!(store, blobId, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:267.

addBlob!(store, entry, linkfile)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:320.

addBlob!(store, blobId, blob, linkfile)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:324.

addBlob!(store, blobId, blob)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:405.

DistributedFactorGraphs.deleteBlob! — Function

Delete a blob from the blob store or dfg with the given entry.

Related deleteBlobEntry!

deleteBlob!(store, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:69.

deleteBlob!(dfg, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:142.

deleteBlob!(store, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:146.

deleteBlob!(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:231.

deleteBlob!(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:282.

deleteBlob!(store, args)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:335.

deleteBlob!(store, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:339.

deleteBlob!(store, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:340.

deleteBlob!(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:420.

DistributedFactorGraphs.getBlob — Function

Get the data blob for the specified blobstore or dfg.

Related getBlobEntry

getBlob(store, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:57.

getBlob(dfg, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:85.

getBlob(store, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:105.

getBlob(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:195.

getBlob(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:263.

getBlob(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:315.

getBlob(store, blobId)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:401.

DistributedFactorGraphs.listBlobs — Function

List all ids in the blob store.

DistributedFactorGraphs.updateBlob! — Function

Update a blob to the blob store or dfg with the given entry. Related updateBlobEntry!

updateBlob!(store, _, _)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:65.

updateBlob!(dfg, entry, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:133.

updateBlob!(store, entry, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:137.

updateBlob!(store, blobId, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:219.

updateBlob!(store, blobId, data)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:275.

updateBlob!(store, blobId, blob)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/BlobStores.jl:413.

DevNotes

TODO TBD update verb on data since data blobs and entries are restricted to immutable only.

DistributedFactorGraphs.packBlob — Function

packBlob

Convert a file (JSON, JPG, PNG, BSON, LAS) to Vector{UInt8} for use as a Blob. Returns the blob and MIME type.

DistributedFactorGraphs.unpackBlob — Function

unpackBlob

Convert a Blob back to the origanal typ using the MIME type or DataFormat type.

DistributedFactorGraphs.addData! — Function

Add both a BlobEntry and Blob to a distributed factor graph or BlobStore. Related addBlobEntry!

addData!(dfg, label, entry, blob; hashfunction, checkhash)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:124.

addData!(
    dfg,
    blobstore,
    label,
    entry,
    blob;
    hashfunction,
    checkhash
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:138.

addData!(
    dfg,
    blobstorekey,
    vLbl,
    bLbl,
    blob,
    timestamp;
    kwargs...
)
addData!(dfg, blobstorekey, vLbl, bLbl, blob; ...)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:153.

addData!(
    dfg,
    blobstore,
    vLbl,
    bLbl,
    blob,
    timestamp;
    description,
    metadata,
    mimeType,
    id,
    blobId,
    originId,
    hashfunction
)
addData!(dfg, blobstore, vLbl, bLbl, blob; ...)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:173.

addData!(
    dfg,
    blobstore,
    vLbl,
    blobLabel,
    blob,
    timestamp;
    description,
    metadata,
    mimeType,
    origin
)
addData!(dfg, blobstore, vLbl, blobLabel, blob; ...)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:206.

DistributedFactorGraphs.deleteData! — Function

Delete a blob entry and blob from the blob store or dfg. Related deleteBlobEntry!

deleteData!(dfg, vLbl, bLbl)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:278.

deleteData!(dfg, blobstore, vLbl, entry)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:284.

deleteData!(dfg, blobstore, vLbl, bLbl)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:293.

DistributedFactorGraphs.getData — Function

Get the blob entry and blob for the specified blobstore or dfg retured as a tuple. Related getBlobEntry

getData(dfg, vlabel, key; hashfunction, checkhash, getlast)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:81.

getData(
    dfg,
    blobstore,
    label,
    key;
    hashfunction,
    checkhash,
    getlast
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:108.

DistributedFactorGraphs.updateData! — Function

Update a blob entry or blob to the blob store or dfg. Related updateBlobEntry!

updateData!(
    dfg,
    label,
    entry,
    blob;
    hashfunction,
    checkhash
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:240.

updateData!(
    dfg,
    blobstore,
    label,
    entry,
    blob;
    hashfunction
)

defined at /home/runner/work/DistributedFactorGraphs.jl/DistributedFactorGraphs.jl/src/DataBlobs/services/HelpersDataWrapEntryBlob.jl:255.