Reference

Function Reference

Function Reference

Graph Types

Creating DFG Factor Graphs

Getting Factor Graph Nodes

Updating Factor Graph Nodes

Deleting Factor Graph Nodes

Getting Adjacency Matrix

Validating Factor Graph Connectivity

Copying Subgraphs

Summaries

Visualization and Plotting

Graph Types

Missing docstring.

Missing docstring for GraphsDFG. Check Documenter's build log for details.

Creating DFG Factor Graphs

DistributedFactorGraphs.addVariable! — Function.

Add a DFGVariable to a DFG.

source

DistributedFactorGraphs.addFactor! — Function.

Add a DFGFactor to a DFG.

addFactor!(dfg, factor)

source

addFactor!(dfg, variables, factor)

source

addFactor!(dfg, variableLabels, factor)

source

Getting Factor Graph Nodes

Missing docstring.

Missing docstring for getVariables. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getFactors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getVariable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getFactor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getNeighbors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ls. Check Documenter's build log for details.

Missing docstring.

Missing docstring for lsf. Check Documenter's build log for details.

Updating Factor Graph Nodes

Missing docstring.

Missing docstring for updateVariable!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for updateFactor!. Check Documenter's build log for details.

Deleting Factor Graph Nodes

Missing docstring.

Missing docstring for deleteVariable!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for deleteFactor!. Check Documenter's build log for details.

Getting Adjacency Matrix

Missing docstring.

Missing docstring for getIncidenceMatrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getIncidenceMatrixSparse. Check Documenter's build log for details.

Validating Factor Graph Connectivity

DistributedFactorGraphs.isFullyConnected — Function.

Checks if the graph is fully connected, returns true if so.

source

DistributedFactorGraphs.hasOrphans — Function.

Checks if the graph is not fully connected, returns true if it is not contiguous.

source

Copying Subgraphs

Missing docstring.

Missing docstring for getSubgraphAroundNode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getSubgraph. Check Documenter's build log for details.

Summaries

DistributedFactorGraphs.getSummary — Function.

Get a summary of the graph (first-class citizens of variables and factors). Returns a AbstractDFGSummary.

source

DistributedFactorGraphs.getSummaryGraph — Function.

Get a summary graph (first-class citizens of variables and factors) with the same structure as the original graph. Note this is a copy of the original. Returns a LightDFG{NoSolverParams, DFGVariableSummary, DFGFactorSummary}.

source

Visualization and Plotting

Missing docstring.

Missing docstring for toDot. Check Documenter's build log for details.

Missing docstring.

Missing docstring for toDotFile. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dfgplot. Check Documenter's build log for details.