Reference

Function Reference

Function Reference

DistributedFactorGraphs.DistributedFactorGraphs — Module.

DistributedFactorGraphs.jl provides a flexible factor graph API for use in the Caesar.jl ecosystem.

The package supplies:

A standardized API for interacting with factor graphs

Implementations of the API for in-memory and database-driven operation

Visualization extensions to validate the underlying graph

source

DistributedFactorGraphs.AbstractBigDataEntry — Type.

abstract type AbstractBigDataEntry

Abstract parent struct for big data entry.

source

DistributedFactorGraphs.AbstractDFG — Type.

abstract type AbstractDFG

Abstract parent struct for a DFG graph.

source

DistributedFactorGraphs.AbstractDFGFactor — Type.

abstract type AbstractDFGFactor <: DFGNode

An abstract DFG factor.

source

DistributedFactorGraphs.AbstractDFGVariable — Type.

abstract type AbstractDFGVariable <: DFGNode

An abstract DFG variable.

source

DistributedFactorGraphs.AbstractDataStore — Type.

abstract type AbstractDataStore

Superclass of all key-value datastores.

source

DistributedFactorGraphs.AbstractParams — Type.

abstract type AbstractParams

Abstract parent struct for solver parameters.

source

DistributedFactorGraphs.CloudGraphsDFG — Method.

Create a new CloudGraphs-based DFG factor graph using a Neo4j.Connection.

source

DistributedFactorGraphs.CloudGraphsDFG — Method.

Create a new CloudGraphs-based DFG factor graph by specifying the Neo4j connection information.

source

DistributedFactorGraphs.DFGFactor — Method.

Construct a DFG factor given a label.

source

DistributedFactorGraphs.DFGNode — Type.

abstract type DFGNode

Abstract parent struct for DFG variables and factors.

source

DistributedFactorGraphs.DFGNodeParams — Type.

mutable struct DFGNodeParams

The common node parameters for variables and factors.

source

DistributedFactorGraphs.DFGVariable — Method.

DFGVariable(label, softtype; timestamp, tags, estimateDict, solverDataDict, smallData, bigData, solvable)

The default DFGVariable constructor.

source

DistributedFactorGraphs.FileBigDataEntry — Type.

struct FileBigDataEntry <: AbstractBigDataEntry

BigDataEntry in a file.

source

DistributedFactorGraphs.FileDataStore — Type.

struct FileDataStore <: AbstractDataStore{UInt8}

Simple file data store with a specified data type and a specified key type.

source

DistributedFactorGraphs.GeneralBigDataEntry — Type.

mutable struct GeneralBigDataEntry <: AbstractBigDataEntry

GeneralBigDataEntry is a generic multipurpose data entry that creates a unique reproducible key using userIdrobotIdsessionIdvariableIdkey.

source

DistributedFactorGraphs.GenericFunctionNodeData — Type.

mutable struct GenericFunctionNodeData{T<:Union{FactorOperationalMemory, FunctorInferenceType, PackedInferenceType}}

Notes

S::Symbol

Designing (WIP)

T <: Union{FactorOperationalMemory, PackedInferenceType}

in IIF.CCW{T <: DFG.InferenceType}

in IIF.FunctorPairwiseMinimize <: InferenceType # DFG whatever, something, we'll figure it out

in Main/User, SomeFactor <: FunctorPairwiseMinimize

source

DistributedFactorGraphs.InMemoryDataStore — Type.

struct InMemoryDataStore{T, E<:AbstractBigDataEntry} <: AbstractDataStore{T}

Simple in-memory data store with a specified data type and a specified key type.

source

DistributedFactorGraphs.InMemoryDataStore — Method.

InMemoryDataStore()

Create an in-memory store using binary data (UInt8) as a type.

source

DistributedFactorGraphs.InMemoryDataStore — Method.

InMemoryDataStore()

Create an in-memory store using a specific data type.

source

DistributedFactorGraphs.MeanMaxPPE — Type.

struct MeanMaxPPE <: AbstractPointParametricEst

Data container to store Parameteric Point Estimate (PPE) for mean and max.

source

DistributedFactorGraphs.MongodbBigDataEntry — Type.

struct MongodbBigDataEntry <: AbstractBigDataEntry

BigDataEntry in MongoDB.

source

DistributedFactorGraphs.NoSolverParams — Type.

struct NoSolverParams <: AbstractParams

Empty structure for solver parameters.

source

DistributedFactorGraphs.PackedVariableNodeData — Type.

mutable struct PackedVariableNodeData

Packed VariabeNodeData structure for serializing DFGVariables.

Fields:

vecval::Array{Float64,1}

dimval::Int64

vecbw::Array{Float64,1}

dimbw::Int64

BayesNetOutVertIDs::Array{Symbol,1}

dimIDs::Array{Int64,1}

dims::Int64

eliminated::Bool

BayesNetVertID::Symbol

separator::Array{Symbol,1}

softtype::String

initialized::Bool

inferdim::Float64

ismargin::Bool

dontmargin::Bool

solveInProgress::Int64

solvedCount::Int64

source

DistributedFactorGraphs.VariableNodeData — Type.

mutable struct VariableNodeData{T<:InferenceVariable}

Data container for solver-specific data.

Fields:

val::Array{Float64,2}

bw::Array{Float64,2}

BayesNetOutVertIDs::Array{Symbol,1}

dimIDs::Array{Int64,1}

dims::Int64

eliminated::Bool

BayesNetVertID::Symbol

separator::Array{Symbol,1}

softtype::InferenceVariable

initialized::Bool

inferdim::Float64

ismargin::Bool

dontmargin::Bool

solveInProgress::Int64

solvedCount::Int64

events::Dict{Symbol,Base.GenericCondition{ReentrantLock}}

source

DistributedFactorGraphs._getDuplicatedEmptyDFG — Method.

_getDuplicatedEmptyDFG(dfg)

Gets an empty and unique CloudGraphsDFG derived from an existing DFG.

source

DistributedFactorGraphs._getDuplicatedEmptyDFG — Method.

Gets an empty and unique CloudGraphsDFG derived from an existing DFG.

source

DistributedFactorGraphs.addBigData! — Method.

Adds the data to the store with the given entry. The function will warn if the entry already exists and will overwrite it.

source

DistributedFactorGraphs.addBigDataEntry! — Method.

addBigDataEntry!(dfg, label, bde)

Add Big Data Entry to distributed factor graph. Should be extended if DFG variable is not returned by reference.

source

DistributedFactorGraphs.addBigDataEntry! — Method.

addBigDataEntry!(var, bde)

Add Big Data Entry to a DFG variable

source

DistributedFactorGraphs.addDataEntry! — Method.

addDataEntry!(dfg, lbl, datastore, descr, mimeType, data)

Add Big Data Entry to distributed factor graph. Should be extended if DFG variable is not returned by reference.

Example

See docs for getDataEntryElement.

Related

addData!, getDataEntryElement, fetchData

source

DistributedFactorGraphs.addFactor! — Method.

addFactor!(dfg, variables, factor)

source

DistributedFactorGraphs.addFactor! — Method.

addFactor!(dfg, variableLabels, factor)

source

DistributedFactorGraphs.addFactor! — Method.

Add a DFGFactor to a DFG.

addFactor!(dfg, factor)

source

DistributedFactorGraphs.addPPE! — Function.

addPPE!(dfg, sourceVariable)
addPPE!(dfg, sourceVariable, ppekey)

Add a new PPE entry from a deepcopy of the source variable PPE. NOTE: Copies the solver data.

source

DistributedFactorGraphs.addPPE! — Method.

addPPE!(dfg, variablekey, ppe)
addPPE!(dfg, variablekey, ppe, ppekey)

Add variable PPE, errors if it already exists.

source

DistributedFactorGraphs.addVariable! — Method.

addVariable!(dfg, variable)

Add a DFGVariable to a DFG.

source

DistributedFactorGraphs.addVariableSolverData! — Function.

addVariableSolverData!(dfg, variablekey, vnd)
addVariableSolverData!(dfg, variablekey, vnd, solvekey)

Add variable solver data, errors if it already exists.

source

DistributedFactorGraphs.addVariableSolverData! — Function.

addVariableSolverData!(dfg, sourceVariable)
addVariableSolverData!(dfg, sourceVariable, solvekey)

Add a new solver data entry from a deepcopy of the source variable solver data. NOTE: Copies the solver data.

source

DistributedFactorGraphs.buildSubgraph — Method.

buildSubgraph(, dfg, variableFactorLabels)
buildSubgraph(, dfg, variableFactorLabels, distance; solvable, sessionId, kwargs...)

Build a deep subgraph copy from the DFG given a list of variables and factors and an optional distance. Note: Orphaned factors (where the subgraph does not contain all the related variables) are not returned. Related:

copyGraph!

getNeighborhood

deepcopyGraph

mergeGraph!

Dev Notes

Bulk vs node for node: a list of labels are compiled and the sugraph is copied in bulk.

source

DistributedFactorGraphs.clearRobot!! — Method.

clearRobot!!(dfg)

DANGER: Clears the whole robot + sessions from the database.

source

DistributedFactorGraphs.clearSession!! — Method.

clearSession!!(dfg)

DANGER: Clears the whole session from the database.

source

DistributedFactorGraphs.clearUser!! — Method.

clearUser!!(dfg)

DANGER: Clears the whole user + robot + sessions from the database.

source

DistributedFactorGraphs.compareAll — Method.

compareAll(Al, Bl; show, skip)

Recursively compare the all fields of T that are not in skip for objects Al and Bl.

TODO > add to func_ref.md

source

DistributedFactorGraphs.compareAllVariables — Method.

Compare all variables in both ::FactorGraphs A and B.

Notes

A and B should all the same variables and factors.

Related:

compareFactorGraphs, compareSimilarVariables, compareVariable, ls

source

DistributedFactorGraphs.compareFactor — Method.

compareFactor(A, B; show, skip, skipsamples, skipcompute)

Compare that all fields are the same in a ::FactorGraph factor.

source

DistributedFactorGraphs.compareFactorGraphs — Method.

Compare and return if two factor graph objects are the same, by comparing similar variables and factors.

Notes:

Default items to skip with skipsamples, skipcompute.

User defined fields to skip can be specified with skip::Vector{Symbol}.

To enable debug messages for viewing which fields are not the same:

https://stackoverflow.com/questions/53548681/how-to-enable-debugging-messages-in-juno-julia-editor

Related:

compareSimilarVariables, compareSimilarFactors, compareAllVariables, ls.

source

DistributedFactorGraphs.compareFields — Method.

compareFields(Al, Bl; show, skip)

Compare the all fields of T that are not in skip for objects Al and Bl.

TODO > add to func_ref.md

source

DistributedFactorGraphs.compareSimilarFactors — Method.

Compare similar factors between ::FactorGraphs A and B.

Related:

compareFactorGraphs, compareSimilarVariables, compareAllVariables, ls.

source

DistributedFactorGraphs.compareSimilarVariables — Method.

compareSimilarVariables(fgA, fgB; skip, show, skipsamples)

Compare similar labels between ::FactorGraphs A and B.

Notes

At least one variable label should exist in both A and B.

Related:

compareFactorGraphs, compareAllVariables, compareSimilarFactors, compareVariable, ls.

source

DistributedFactorGraphs.compareVariable — Method.

compareVariable(A, B; skip, show, skipsamples)

Compare that all fields are the same in a ::FactorGraph variable.

source

DistributedFactorGraphs.copyGraph! — Method.

copyGraph!(destDFG, sourceDFG, variableLabels, factorLabels; copyGraphMetadata, overwriteDest, deepcopyNodes, verbose)

Common function for copying nodes from one graph into another graph. This is overridden in specialized implementations for performance. Orphaned factors are not added, with a warning if verbose. Set overwriteDest to overwrite existing variables and factors in the destination DFG. NOTE: copyGraphMetadata not supported yet. Related:

deepcopyGraph

deepcopyGraph!

buildSubgraph

getNeighborhood

mergeGraph!

Dev Note:

Adapted from and replaces internal _copyIntoGraph!

source

DistributedFactorGraphs.copySession! — Method.

copySession!(sourceDFG, destDFG)

DANGER: Copies and overwrites the destination session. If no destination specified then it creates a unique one.

source

DistributedFactorGraphs.copySession! — Method.

copySession!(sourceDFG)

DANGER: Copies the source to a new unique destination.

source

DistributedFactorGraphs.copyStore — Method.

copyStore(sourceStore, destStore; sourceEntries)

Copies all the entries from the source into the destination. Can specify which entries to copy with the sourceEntries parameter. Returns the list of copied entries.

source

DistributedFactorGraphs.createDfgSessionIfNotExist — Method.

createDfgSessionIfNotExist(dfg)

Shortcut method to create the user, robot, and session if it doesn't already exist.

source

DistributedFactorGraphs.deleteBigData! — Method.

Delete the data in the store for the given entry. The function will error and return nothing if the entry does not exist.

source

DistributedFactorGraphs.deleteBigDataEntry! — Method.

deleteBigDataEntry!(var, key)

Delete big data entry from the factor graph. Note this doesn't remove it from any data stores.

Notes:

users responsibility to delete big data in db before deleting entry

source

DistributedFactorGraphs.deleteFactor! — Method.

deleteFactor!(dfg, factor)

Delete the referened DFGFactor from the DFG.

source

DistributedFactorGraphs.deleteFactor! — Method.

Delete a DFGFactor from the DFG using its label.

source

DistributedFactorGraphs.deletePPE! — Function.

deletePPE!(dfg, sourceVariable)
deletePPE!(dfg, sourceVariable, ppekey)

Delete PPE data, returns the deleted element.

source

DistributedFactorGraphs.deletePPE! — Function.

deletePPE!(dfg, variablekey)
deletePPE!(dfg, variablekey, ppekey)

Delete PPE data, returns the deleted element.

source

DistributedFactorGraphs.deleteVariable! — Method.

deleteVariable!(dfg, variable)

Delete a referenced DFGVariable from the DFG.

source

DistributedFactorGraphs.deleteVariable! — Method.

deleteVariable!(dfg, label)

Delete a DFGVariable from the DFG using its label.

source

DistributedFactorGraphs.deleteVariableSolverData! — Function.

deleteVariableSolverData!(dfg, sourceVariable)
deleteVariableSolverData!(dfg, sourceVariable, solvekey)

Delete variable solver data, returns the deleted element.

source

DistributedFactorGraphs.deleteVariableSolverData! — Function.

deleteVariableSolverData!(dfg, variablekey)
deleteVariableSolverData!(dfg, variablekey, solvekey)

Delete variable solver data, returns the deleted element.

source

DistributedFactorGraphs.emptyTags! — Method.

emptyTags!(dfg, sym)

Empty all tags from the node (empty)

source

DistributedFactorGraphs.exists — Method.

True if the variable or factor exists in the graph.

source

DistributedFactorGraphs.findClosestTimestamp — Method.

findClosestTimestamp(setA, setB)

Find and return the closest timestamp from two sets of Tuples. Also return the minimum delta-time (::Millisecond) and how many elements match from the two sets are separated by the minimum delta-time.

source

DistributedFactorGraphs.findVariableNearTimestamp — Function.

findVariableNearTimestamp(dfg, timest)
findVariableNearTimestamp(dfg, timest, regexFilter; tags, solvable, warnDuplicate, number)

Find and return nearest variable labels per delta time. Function will filter on regexFilter, tags, and solvable.

DevNotes:

TODO number should allow returning more than one for k-nearest matches.

Future versions likely will require some optimization around the internal getVariable call.

Perhaps a dedicated/efficient getVariableTimestamp for all DFG flavors.

Related

ls, listVariables, findClosestTimestamp

source

DistributedFactorGraphs.getAddHistory — Method.

getAddHistory(dfg)

source

DistributedFactorGraphs.getBiadjacencyMatrix — Method.

getBiadjacencyMatrix(dfg; solvable)

Get a matrix indicating adjacency between variables and factors. Returned as a named tuple: B::SparseMatrixCSC{Int}, varLabels::Vector{Symbol) facLabels::Vector{Symbol). Rows are the factors, columns are the variables, with the corresponding labels in varLabels,facLabels.

source

DistributedFactorGraphs.getBigData — Method.

getBigData(store, entry)

Get the data for the specified entry, returns the data or Nothing.

source

DistributedFactorGraphs.getBigDataEntries — Method.

getBigDataEntries(var)

Get big data entries, Vector{AbstractBigDataEntry}

source

DistributedFactorGraphs.getBigDataEntry — Method.

getBigDataEntry(var, key)

Get big data entry

source

DistributedFactorGraphs.getBigDataKeys — Method.

getBigDataKeys(var)

getBigDataKeys

source

DistributedFactorGraphs.getDFGInfo — Method.

getDFGInfo(dfg)

Convenience function to get all the metadata of a DFG

source

DistributedFactorGraphs.getDataEntryElement — Method.

getDataEntryElement(dfg, dfglabel, datastore, datalabel)

Get both the entry and raw data element from datastore returning as a tuple.

Notes:

This is the counterpart to addDataEntry!.

Data is identified by the node in the DFG object dfglabel::Symbol as well as datalabel::Symbol.

The data should have been stored along with a entry.mimeType::String which describes the format of the data.

ImageMagick.jl is useful for storing images in png or jpg compressed format.

Example

# some dfg object
fg = initfg()
addVariable!(fg, :x0, IIF.ContinuousScalar) # using IncrementalInference
# FileDataStore (as example)
datastore = FileDataStore(joinLogPath(fg,"datastore"))

# now some data comes in
mydata = Dict(:soundBite => randn(Float32, 10000), :meta => "something about lazy foxes and fences.")

# store the data
addDataEntry!( fg, :x0, datastore, :SOUND_DATA, "application/json", Vector{UInt8}(JSON2.write( mydata ))  )

# get/fetch the data
entry, rawData = fetchData(fg, :x0, datastore, :SOUND_DATA)

# unpack data to original format (this must be done by the user)
@show entry.mimeType # "applicatio/json"
userFormat = JSON2.read(IOBuffer(rawData))

Related

addDataEntry!, addData!, fetchData, fetchDataEntryElement

source

DistributedFactorGraphs.getDescription — Method.

getDescription(dfg)

source

DistributedFactorGraphs.getFactor — Method.

Get a DFGFactor from a DFG using its label.

source

DistributedFactorGraphs.getFactorFunction — Method.

getFactorFunction(fcd)

Return reference to the user factor in <:AbstractDFG identified by ::Symbol.

source

DistributedFactorGraphs.getFactorType — Method.

getFactorType(data)

Return user factor type from factor graph identified by label ::Symbol.

Notes

Replaces older getfnctype.

source

DistributedFactorGraphs.getFactors — Method.

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

DistributedFactorGraphs.getLabel — Method.

getLabel(v)

Return the label for a DFGNode.

source

DistributedFactorGraphs.getLastUpdatedTimestamp — Method.

getLastUpdatedTimestamp(est)

source

DistributedFactorGraphs.getMaxPPE — Method.

getMaxPPE(est)

source

DistributedFactorGraphs.getMeanPPE — Method.

getMeanPPE(est)

source

DistributedFactorGraphs.getNeighborhood — Method.

getNeighborhood(dfg, label, distance)

Build a list of all unique neighbors inside 'distance' Related:

copyGraph!

buildSubgraph

deepcopyGraph

mergeGraph!

source

DistributedFactorGraphs.getNeighbors — Method.

getNeighbors(dfg, label; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.getPPE — Function.

getPPE(dfg, variablekey)
getPPE(dfg, variablekey, ppekey)

Get the parametric point estimate (PPE) for a variable in the factor graph for a given solve key.

Notes

Defaults on keywords solveKey and method

Related getMeanPPE, getMaxPPE, getKDEMean, getKDEFit, getPPEs, getVariablePPEs

source

DistributedFactorGraphs.getPPE — Function.

getPPE(vari)
getPPE(vari, solveKey)

Get the parametric point estimate (PPE) for a variable in the factor graph.

Notes

Defaults on keywords solveKey and method

Related

getMeanPPE, getMaxPPE, getKDEMean, getKDEFit, getPPEs, getVariablePPEs

source

DistributedFactorGraphs.getPPEDict — Method.

getPPEDict(v)

Get the PPE dictionary for a variable. Recommended to use CRUD operations instead, getPPE, addPPE!, updatePPE!, deletePPE!.

source

DistributedFactorGraphs.getRobot — Method.

getRobot(dfg, userId, robotId)

Get a robot specified by userId:robotId. Returns nothing if it isn't found.

source

DistributedFactorGraphs.getRobot — Method.

getRobot(dfg)

Get the robot specified by the DFG object. Returns nothing if it isn't found.

source

DistributedFactorGraphs.getRobotData — Method.

getRobotData(dfg)

Get the robot data associated with the graph.

source

DistributedFactorGraphs.getRobotId — Method.

getRobotId(dfg)

source

DistributedFactorGraphs.getSession — Method.

getSession(dfg, userId, robotId, sessionId)

Get a session specified by userId:robotId:sessionId. Returns nothing if it isn't found.

source

DistributedFactorGraphs.getSession — Method.

getSession(dfg)

Get the session specified by the DFG object. Returns nothing if it isn't found.

source

DistributedFactorGraphs.getSessionData — Method.

getSessionData(dfg)

Get the session data associated with the graph.

source

DistributedFactorGraphs.getSessionId — Method.

getSessionId(dfg)

source

DistributedFactorGraphs.getSmallData — Method.

getSmallData(v)

Get the small data for a variable.

source

DistributedFactorGraphs.getSofttype — Method.

getSofttype(vnd)

Variable nodes softtype information holding a variety of meta data associated with the type of variable stored in that node of the factor graph.

Related

getVariableType

source

DistributedFactorGraphs.getSofttypename — Method.

getSofttypename(v)

Retrieve the soft type name symbol for a DFGVariableSummary. ie :Point2, Pose2, etc.

source

DistributedFactorGraphs.getSolvable — Method.

getSolvable(dfg, sym)

Get 'solvable' parameter for either a variable or factor.

source

DistributedFactorGraphs.getSolvable — Method.

getSolvable(var)

Variables or factors may or may not be 'solvable', depending on a user definition. Useful for ensuring atomic transactions.

Related:

isSolveInProgress

source

DistributedFactorGraphs.getSolveInProgress — Function.

getSolveInProgress(var)
getSolveInProgress(var, solveKey)

Which variables or factors are currently being used by an active solver. Useful for ensuring atomic transactions.

DevNotes:

Will be renamed to data.solveinprogress which will be in VND, not DFGNode – see DFG #201

Related

isSolvable

source

DistributedFactorGraphs.getSolvedCount — Method.

getSolvedCount(v)

Get the number of times a variable has been inferred – i.e. solvedCount.

Related

isSolved, setSolvedCount!

source

DistributedFactorGraphs.getSolverData — Function.

getSolverData(v)
getSolverData(v, key)

Retrieve solver data structure stored in a variable.

source

DistributedFactorGraphs.getSolverData — Method.

getSolverData(f)

Retrieve solver data structure stored in a factor.

source

DistributedFactorGraphs.getSolverDataDict — Method.

getSolverDataDict(v)

Get solver data dictionary for a variable. Advised to use graph CRUD operations instead.

source

DistributedFactorGraphs.getSolverParams — Method.

getSolverParams(dfg)

source

DistributedFactorGraphs.getSuggestedPPE — Method.

getSuggestedPPE(est)

source

DistributedFactorGraphs.getSummary — Method.

Get a summary of the graph (first-class citizens of variables and factors). Returns a DFGSummary.

source

DistributedFactorGraphs.getSummaryGraph — Method.

Get a summary graph (first-class citizens of variables and factors) with the same structure as the original graph. Note this is a copy of the original. Returns a LightDFG{NoSolverParams, DFGVariableSummary, DFGFactorSummary}.

source

DistributedFactorGraphs.getTags — Method.

getTags(v)

Return the tags for a DFGNode.

source

DistributedFactorGraphs.getTimestamp — Method.

getTimestamp(v)

Get the timestamp of a DFGNode.

source

DistributedFactorGraphs.getUser — Method.

getUser(dfg, userId)

Get a user specified by userId. Returns nothing if it isn't found.

source

DistributedFactorGraphs.getUser — Method.

getUser(dfg)

Get the user specified by the DFG object. Returns nothing if it isn't found.

source

DistributedFactorGraphs.getUserData — Method.

getUserData(dfg)

Get the user data associated with the graph.

source

DistributedFactorGraphs.getUserId — Method.

getUserId(dfg)

source

DistributedFactorGraphs.getVariable — Method.

getVariable(dfg, label, solveKey)

Get a DFGVariable with a specific solver key. In memory types still return a reference, other types returns a variable with only solveKey.

source

DistributedFactorGraphs.getVariable — Method.

getVariable(dfg, label)

Get a DFGVariable from a DFG using its label.

source

DistributedFactorGraphs.getVariableOrder — Method.

getVariableOrder(fct)

Get the variable ordering for this factor. Should be equivalent to getNeighbors unless something was deleted in the graph.

source

DistributedFactorGraphs.getVariablePPEDict — Method.

getVariablePPEDict(vari)

Return full dictionary of PPEs in a variable, recommended to rather use CRUD: getPPE,

source

DistributedFactorGraphs.getVariableSolverData — Function.

getVariableSolverData(dfg, variablekey)
getVariableSolverData(dfg, variablekey, solvekey)

Get variable solverdata for a given solve key.

source

DistributedFactorGraphs.getVariableType — Method.

getVariableType

Alias for getSofttype.

source

DistributedFactorGraphs.getVariables — Method.

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

source

DistributedFactorGraphs.hasDataEntry — Method.

hasDataEntry(var, key)

Does a data entry (element) exist at key.

source

DistributedFactorGraphs.hasTags — Method.

hasTags(dfg, sym, tags; matchAll)

Determine if the variable or factor neighbors have the tags:;Vector{Symbol}, and matchAll::Bool.

source

DistributedFactorGraphs.hasTagsNeighbors — Method.

hasTagsNeighbors(dfg, sym, tags; matchAll)

Determine if the variable or factor neighbors have the tags:;Vector{Symbol}, and matchAll::Bool.

source

DistributedFactorGraphs.isConnected — Method.

isConnected(dfg)

Checks if the graph is fully connected, returns true if so.

source

DistributedFactorGraphs.isFactor — Method.

Return whether sym::Symbol represents a factor vertex in the graph DFG. Checks whether it both exists in the graph and is a factor. (If you rather want a quicker for type, just do node isa DFGFactor)

source

DistributedFactorGraphs.isInitialized — Function.

isInitialized(var)
isInitialized(var, key)

Returns state of vertex data .initialized flag.

Notes:

used by both factor graph variable and Bayes tree clique logic.

source

DistributedFactorGraphs.isPrior — Method.

Return ::Bool on whether given factor fc::Symbol is a prior in factor graph dfg.

source

DistributedFactorGraphs.isSolvable — Method.

isSolvable(node)

Variables or factors may or may not be 'solvable', depending on a user definition. returns true if getSolvable > 0 Related:

getSolvable(@ref)

source

DistributedFactorGraphs.isSolved — Method.

isSolved(v)

Boolean on whether the variable has been solved.

Related

getSolved, setSolved!

source

DistributedFactorGraphs.isValidLabel — Method.

isValidLabel(id)

Returns true if the label is valid for a session, robot, or user ID.

source

DistributedFactorGraphs.isVariable — Method.

Return whether sym::Symbol represents a variable vertex in the graph DFG. Checks whether it both exists in the graph and is a variable. (If you rather want a quick for type, just do node isa DFGVariable)

source

DistributedFactorGraphs.listFactors — Method.

listFactors(dfg)
listFactors(dfg, regexFilter; tags, solvable)

Get a list of the IDs (labels) of the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

DistributedFactorGraphs.listPPEs — Method.

listPPEs(dfg, variablekey)

List all the PPE data keys in the variable.

source

DistributedFactorGraphs.listStoreEntries — Method.

listStoreEntries(store)

List all entries in the data store.

source

DistributedFactorGraphs.listTags — Method.

listTags(dfg, sym)

Return the tags for a variable or factor.

source

DistributedFactorGraphs.listVariableSolverData — Method.

listVariableSolverData(dfg, variablekey)

List all the solver data keys in the variable.

source

DistributedFactorGraphs.listVariables — Method.

Get a list of IDs of the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

Example

listVariables(dfg, r"l", tags=[:APRILTAG;])

Related:

ls

source

DistributedFactorGraphs.loadDFG! — Method.

loadDFG!(dfgLoadInto, dst)

Load a DFG from a saved folder. Always provide the IIF module as the second parameter.

Example

using DistributedFactorGraphs, IncrementalInference
# Create a DFG - can make one directly, e.g. LightDFG{NoSolverParams}() or use IIF:
dfg = initfg()
# Load the graph
loadDFG("/tmp/savedgraph.tar.gz", IncrementalInference, dfg)
# Use the DFG as you do normally.
ls(dfg)

source

DistributedFactorGraphs.ls — Method.

ls(dfg)
ls(dfg, label; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.ls — Method.

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

source

DistributedFactorGraphs.ls — Method.

ls(dfg)
ls(dfg, node; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

DistributedFactorGraphs.ls2 — Method.

ls2(dfg, label)

Helper to return neighbors at distance 2 around a given node label.

source

DistributedFactorGraphs.lsRobots — Method.

lsRobots(dfg)

List all robots for the specified DFG's user. Returns nothing if it isn't found.

source

DistributedFactorGraphs.lsSessions — Method.

lsSessions(dfg)

List all sessions for the specified DFG's robot and user. Returns nothing if it isn't found.

source

DistributedFactorGraphs.lsTypes — Method.

Return ::Dict{Symbol, Vector{String}} of all unique variable types in factor graph.

source

DistributedFactorGraphs.lsUsers — Method.

lsUsers(dfg)

List all users. Returns nothing if it isn't found.

source

DistributedFactorGraphs.lsWho — Method.

lsWho(dfg, type)

Gives back all factor labels that fit the bill: lsWho(dfg, :Pose3)

Dev Notes

Cloud versions will benefit from less data transfer

ls(dfg::C, ::T) where {C <: CloudDFG, T <: ..}

Related

ls, lsf, lsfPriors

source

DistributedFactorGraphs.lsf — Method.

Alias for getNeighbors - returns neighbors around a given node label.

source

DistributedFactorGraphs.lsf — Method.

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

DistributedFactorGraphs.lsfPriors — Method.

Return vector of prior factor symbol labels in factor graph dfg.

source

DistributedFactorGraphs.lsfTypes — Method.

Return ::Dict{Symbol, Vector{String}} of all unique factor types in factor graph.

source

DistributedFactorGraphs.lsfWho — Method.

lsfWho(dfg, type)

Gives back all factor labels that fit the bill: lsfWho(dfg, :Point2Point2)

Dev Notes

Cloud versions will benefit from less data transfer

ls(dfg::C, ::T) where {C <: CloudDFG, T <: ..}

Related

ls, lsf, lsfPriors

source

DistributedFactorGraphs.mergeGraph! — Function.

mergeGraph!(destDFG, sourceDFG)
mergeGraph!(destDFG, sourceDFG, variableLabels)
mergeGraph!(destDFG, sourceDFG, variableLabels, factorLabels)
mergeGraph!(destDFG, sourceDFG, variableLabels, factorLabels, distance; solvable, kwargs...)

Merger sourceDFG to destDFG given an optional list of variables and factors and distance. Notes:

Nodes already in the destination graph are updated from sourceDFG.

Orphaned factors (where the subgraph does not contain all the related variables) are not included.

Related:

copyGraph!

buildSubgraph

getNeighborhood

deepcopyGraph

source

DistributedFactorGraphs.mergeGraphVariableData! — Method.

mergeGraphVariableData!(destDFG, sourceDFG, varSyms)

Common function to update all solver data and estimates from one graph to another. This should be used to push local solve data back into a cloud graph, for example.

source

DistributedFactorGraphs.mergePPEs! — Method.

mergePPEs!(destVariable, sourceVariable)

Merges and updates solver and estimate data for a variable (variable can be from another graph). Note: Makes a copy of the estimates and solver data so that there is no coupling between graphs.

source

DistributedFactorGraphs.mergeTags! — Method.

mergeTags!(dfg, sym, tags)

Merge add tags to a variable or factor (union)

source

DistributedFactorGraphs.mergeVariableData! — Method.

mergeVariableData!(dfg, sourceVariable)

Merges and updates solver and estimate data for a variable (variable can be from another graph). Note: Makes a copy of the estimates and solver data so that there is no coupling between graphs.

source

DistributedFactorGraphs.mergeVariableSolverData! — Method.

mergeVariableSolverData!(destVariable, sourceVariable)

Merges and updates solver and estimate data for a variable (variable can be from another graph). If the same key is present in another collection, the value for that key will be the value it has in the last collection listed (updated). Note: Makes a copy of the estimates and solver data so that there is no coupling between graphs.

source

DistributedFactorGraphs.printFactor — Method.

printFactor(dfg, sym)

Display and return to console the user factor identified by tag name.

source

DistributedFactorGraphs.printVariable — Method.

printVariable(dfg, sym)

Display the content of VariableNodeData to console for a given factor graph and variable tag::Symbol.

Dev Notes

TODO split as two show macros between AMP and DFG

source

DistributedFactorGraphs.removeTags! — Method.

removeTags!(dfg, sym, tags)

Remove the tags from the node (setdiff)

source

DistributedFactorGraphs.saveDFG — Method.

saveDFG(dfg, folder)

Save a DFG to a folder. Will create/overwrite folder if it exists.

DevNotes:

TODO remove compress kwarg.

Example

using DistributedFactorGraphs, IncrementalInference
# Create a DFG - can make one directly, e.g. LightDFG{NoSolverParams}() or use IIF:
dfg = initfg()
# ... Add stuff to graph using either IIF or DFG:
v1 = addVariable!(dfg, :a, ContinuousScalar, labels = [:POSE], solvable=0)
# Now save it:
saveDFG(dfg, "/tmp/saveDFG.tar.gz")

source

DistributedFactorGraphs.setDescription! — Method.

setDescription!(dfg, description)

source

DistributedFactorGraphs.setRobotData! — Method.

setRobotData!(dfg, data)

Set the robot data associated with the graph.

source

DistributedFactorGraphs.setSerializationModule! — Method.

setSerializationModule!(dfg, mod)

Deserialization of IncrementalInference objects require discovery of foreign types.

Example:

Template to tunnel types from a user module:

# or more generic solution -- will always try Main if available
IIF.setSerializationNamespace!(Main)

# or a specific package such as RoME if you import all variable and factor types into a specific module.
using RoME
IIF.setSerializationNamespace!(RoME)

source

DistributedFactorGraphs.setSessionData! — Method.

setSessionData!(dfg, data)

Set the session data associated with the graph.

source

DistributedFactorGraphs.setSmallData! — Method.

setSmallData!(v, smallData)

Set the small data for a variable. This will overwrite old smallData.

source

DistributedFactorGraphs.setSolvable! — Method.

setSolvable!(dfg, sym, solvable)

Set the solvable parameter for either a variable or factor.

source

DistributedFactorGraphs.setSolvable! — Method.

setSolvable!(node, solvable)

Set the solvable parameter for either a variable or factor.

source

DistributedFactorGraphs.setSolvedCount! — Method.

setSolvedCount!(v, val)

Update/set the solveCount value.

Related

getSolved, isSolved

source

DistributedFactorGraphs.setSolverData! — Function.

setSolverData!(v, data)
setSolverData!(v, data, key)

Set solver data structure stored in a variable.

source

DistributedFactorGraphs.setTags! — Method.

setTags!(f, tags)

Set the tags for a DFGNode.

source

DistributedFactorGraphs.setTimestamp! — Method.

setTimestamp!(dfg, lbl, ts)

Set the timestamp of a Variable/Factor object in a factor graph. Note: Since timestamp is not mutable setTimestamp! calls updateVariable! internally. See also setTimestamp

source

DistributedFactorGraphs.setTimestamp — Method.

setTimestamp(v, ts; verbose)

Set the timestamp of a DFGVariable object returning a new DFGVariable. Note: Since the timestamp field is not mutable setTimestamp returns a new variable with the updated timestamp (note the absence of !). Use updateVariable! on the returened variable to update it in the factor graph if needed. Alternatively use setTimestamp!. See issue #315.

source

DistributedFactorGraphs.setUserData! — Method.

setUserData!(dfg, data)

Set the user data associated with the graph.

source

DistributedFactorGraphs.sortDFG — Method.

sortDFG(vars; by, kwargs...)

Convenience wrapper for Base.sort. Sort variable (factor) lists in a meaningful way (by timestamp, label, etc), for example [:april;:x1_3;:x1_6;] Defaults to sorting by timestamp for variables and factors and using natural_lt for Symbols. See Base.sort for more detail.

Notes

Not fool proof, but does better than native sort.

Example

sortDFG(ls(dfg)) sortDFG(ls(dfg), by=getLabel, lt=natural_lt)

Related

ls, lsf

source

DistributedFactorGraphs.toDot — Method.

toDot(dfg)

Produces a dot-format of the graph for visualization.

source

DistributedFactorGraphs.toDotFile — Function.

toDotFile(dfg)
toDotFile(dfg, fileName)

Produces a dot file of the graph for visualization. Download XDot to see the data

Note

Default location "/tmp/dfg.dot" – MIGHT BE REMOVED

Can be viewed with the xdot system application.

Based on graphviz.org

source

DistributedFactorGraphs.updateBigData! — Method.

updateBigData!(store, entry, data)

Update the data in the store. The function will error and return nothing if the entry does not exist.

source

DistributedFactorGraphs.updateBigDataEntry! — Method.

updateBigDataEntry!(var, bde)

Update big data entry

DevNote

DF, unclear if update verb is applicable in this case, see #404

source

DistributedFactorGraphs.updateFactor! — Method.

Update a complete DFGFactor in the DFG.

source

DistributedFactorGraphs.updatePPE! — Function.

updatePPE!(dfg, sourceVariable)
updatePPE!(dfg, sourceVariable, ppekey)

Update PPE data if it exists, otherwise add it. NOTE: Copies the PPE data.

source

DistributedFactorGraphs.updatePPE! — Function.

updatePPE!(dfg, sourceVariables)
updatePPE!(dfg, sourceVariables, ppekey)

Update PPE data if it exists, otherwise add it.

source

DistributedFactorGraphs.updatePPE! — Method.

Update PPE data if it exists, otherwise add it – one call per key::Symbol=:default.

source

DistributedFactorGraphs.updateVariable! — Method.

Update a complete DFGVariable in the DFG.

source

DistributedFactorGraphs.updateVariableSolverData! — Function.

updateVariableSolverData!(dfg, variablekey, vnd)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey, useCopy)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey, useCopy, fields)
updateVariableSolverData!(dfg, variablekey, vnd, solvekey, useCopy, fields, verbose)

Update variable solver data if it exists, otherwise add it.

Notes:

useCopy=true to copy solver data and keep separate memory.

Use fields to updated only a few VND.fields while adhering to useCopy.

Related

mergeVariableSolverData!

source

DistributedFactorGraphs._bindSessionNodeToInitialVariable — Method.

_bindSessionNodeToInitialVariable(neo4jInstance, userId, robotId, sessionId, initialVariableLabel)

Bind the SESSION node to the inital variable. Checks for existence.

source

DistributedFactorGraphs._createNode — Function.

_createNode(neo4jInstance, labels, properties, parentNode)
_createNode(neo4jInstance, labels, properties, parentNode, relationshipLabel)

Create a node and optionally specify a parent. Note: Using symbols so that the labels obey Neo4j requirements

source

DistributedFactorGraphs._dictToNeo4jProps — Method.

_dictToNeo4jProps(dict)

Build a Cypher-compliant set of properies from a JSON dictionary. Note individual values are serialized if they are not already.

source

DistributedFactorGraphs._getLabelsForInst — Method.

_getLabelsForInst(dfg, inst; parentKey)

Get the Neo4j labels for any node instance.

source

DistributedFactorGraphs._getLabelsForType — Method.

_getLabelsForType(dfg, type; parentKey)

Get the Neo4j labels for any node type.

source

DistributedFactorGraphs._getLabelsFromCyphonQuery — Function.

_getLabelsFromCyphonQuery(neo4jInstance, matchCondition)
_getLabelsFromCyphonQuery(neo4jInstance, matchCondition, orderProperty)

Returns the list of CloudGraph nodes that matches the Cyphon query. The nodes of interest should be labelled 'node' because the query will use the return of id(node) #Example matchCondition = '(u:User)-[:ROBOT]->(n:Robot:NewRobot)-[:SESSION]->(node:Session)'. Will return all sessions because of 'node:Session'. If orderProperty is not ==, then 'order by n.{orderProperty} will be appended to the query'. So can make orderProperty = label or id.

source

DistributedFactorGraphs._getNeoNodesFromCyphonQuery — Function.

_getNeoNodesFromCyphonQuery(neo4jInstance, matchCondition)
_getNeoNodesFromCyphonQuery(neo4jInstance, matchCondition, orderProperty)

Returns the list of CloudGraph nodes that matches the Cyphon query. The nodes of interest should be labelled 'node' because the query will use the return of id(node) #Example matchCondition = '(u:User)-[:ROBOT]->(n:Robot:NewRobot)-[:SESSION]->(node:Session)'. Will return all sessions because of 'node:Session'. If orderProperty is not ==, then 'order by n.{orderProperty} will be appended to the query'. So can make orderProperty = label or id.

source

DistributedFactorGraphs._getNodeProperty — Method.

_getNodeProperty(neo4jInstance, nodeLabels, property)

Get a node property - returns nothing if not found

source

DistributedFactorGraphs._getNodeTags — Method.

_getNodeTags(neo4jInstance, nodeLabels)

Get a node's tags

source

DistributedFactorGraphs._getSessionNeoNodesForRobot — Method.

_getSessionNeoNodesForRobot(neo4jInstance, userId, robotId)

Utility function to get a Neo4j node sessions for a robot.

source

DistributedFactorGraphs._queryNeo4j — Method.

_queryNeo4j(neo4jInstance, query)

Returns the transaction for a given query. NOTE: Must commit(transaction) after you're done.

source

DistributedFactorGraphs._setNodeProperty — Method.

_setNodeProperty(neo4jInstance, nodeLabels, property, value)

Set a node property - returns count of changed nodes.

source

DistributedFactorGraphs._tryGetNeoNodeIdFromNodeLabel — Method.

_tryGetNeoNodeIdFromNodeLabel(neo4jInstance, userId, robotId, sessionId, nodeLabel; nodeType)

Try get a Neo4j node ID from a node label.

source

DistributedFactorGraphs._uniqueKey — Method.

_uniqueKey(dfg, v, key)

Internal function to generate a unique key for the entry - userIdrobotIdsessionIdvariablekey. Simple symbol.

source

DistributedFactorGraphs.getAdjacencyMatrixSymbols — Method.

getAdjacencyMatrixSymbols(dfg; solvable)

Get a matrix indicating relationships between variables and factors. Rows are all factors, columns are all variables, and each cell contains either nothing or the symbol of the relating factor. The first row and first column are factor and variable headings respectively. Note: rather use getBiadjacencyMatrix

source

DistributedFactorGraphs.getFactorOperationalMemoryType — Method.

getFactorOperationalMemoryType(dummy)

Method must be overloaded by the user for Serialization to work. E.g. IncrementalInference uses CommonConvWrapper <: FactorOperationalMemory.

source

DistributedFactorGraphs.getTypeFromSerializationModule — Method.

Get a type from the serialization module inside DFG.

source

DistributedFactorGraphs.packPPE — Method.

packPPE(dfg, ppe)

Pack a PPE into a Dict{String, Any}.

source

DistributedFactorGraphs.rebuildFactorMetadata! — Method.

rebuildFactorMetadata!(dfg, factor)

Method must be overloaded by the user for Serialization to work.

source

DistributedFactorGraphs.setRobotId! — Method.

setRobotId!(dfg, robotId)

source

DistributedFactorGraphs.setSessionId! — Method.

setSessionId!(dfg, sessionId)

source

DistributedFactorGraphs.setUserId! — Method.

setUserId!(dfg, userId)

source

DistributedFactorGraphs.unpackPPE — Method.

unpackPPE(dfg, packedPPE)

Unpack a Dict{String, Any} into a PPE.

source

DistributedFactorGraphs.LightDFGs.LightDFG — Type.

An in-memory DistributedFactorGraph based on LightGraphs.jl with parameters:

T: Solver parameters (defaults to NoSolverParams())

V: Variable type

F: Factor type

source

DistributedFactorGraphs.LightDFGs.LightDFG — Method.

LightDFG()
LightDFG(g; params, kwargs...)

Create an in-memory LightDFG with the following parameters:

T: Solver parameters (defaults to NoSolverParams())

V: Variable type

F: Factor type

source

DistributedFactorGraphs.LightDFGs.LightDFG — Method.

LightDFG()

Create an in-memory LightDFG with the following parameters:

T: Solver parameters (defaults to NoSolverParams())

V: Variable type

F: Factor type

source

DistributedFactorGraphs.LightDFGs.savedot_attributes — Method.

savedot_attributes(io, dfg)

A replacement for to_dot that saves only hardcoded factor graph plotting attributes.

source

DistributedFactorGraphs._getDuplicatedEmptyDFG — Method.

_getDuplicatedEmptyDFG(dfg)

Gets an empty and unique LightDFG derived from an existing DFG.

source

DistributedFactorGraphs.DFGPlots.dfgplot — Function.

dfgplot(dfg)
dfgplot(dfg, p)

Plots the structure of the factor graph. GraphPlot must be imported before DistributedFactoGraphs for these functions to be available. Returns the plot context.

E.g.

using GraphPlot
using DistributedFactorGraphs, DistributedFactorGraphs.DFGPlots
# ... Make graph...
# Using GraphViz plotting
dfgplot(fg)
# Save to PDF
using Compose
draw(PDF("/tmp/graph.pdf", 16cm, 16cm), dfgplot(fg))

More information at GraphPlot.jl

source

DistributedFactorGraphs.DFGPlots.dfgplot — Function.

dfgplot(dfg)
dfgplot(dfg, p)

Plots the structure of the factor graph. GraphPlot must be imported before DistributedFactorGraphs for these functions to be available. Returns the plot context.

E.g.

using GraphPlot
using DistributedFactorGraphs, DistributedFactorGraphs.DFGPlots
# ... Make graph...
# Using GraphViz plotting
dfgplot(fg)
# Save to PDF
using Compose
draw(PDF("/tmp/graph.pdf", 16cm, 16cm), dfgplot(fg))

More information at GraphPlot.jl

source