Reference

Function Reference

Function Reference

Graph Types

Creating DFG Factor Graphs

Getting Factor Graph Nodes

Updating Factor Graph Nodes

Deleting Factor Graph Nodes

Getting Adjacency Matrix

Validating Factor Graph Connectivity

Copying Subgraphs

DataFrame Extension Functions

Graph Types

DistributedFactorGraphs.GraphsJl.GraphsDFG — Type.

GraphsDFG()

Create a new in-memory Graphs.jl-based DFG factor graph.

source

Creating DFG Factor Graphs

DistributedFactorGraphs.GraphsJl.addVariable! — Function.

addVariable!(dfg, variable)

Add a DFGVariable to a DFG.

source

DistributedFactorGraphs.GraphsJl.addFactor! — Function.

addFactor!(dfg, factor, variables)

Add a DFGFactor to a DFG.

source

Getting Factor Graph Nodes

DistributedFactorGraphs.GraphsJl.getVariables — Function.

getVariables(dfg)
getVariables(dfg, regexFilter)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables.

source

DistributedFactorGraphs.GraphsJl.getFactors — Function.

getFactors(dfg)
getFactors(dfg, regexFilter)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

DistributedFactorGraphs.GraphsJl.getVariable — Function.

getVariable(dfg, variableId)

Get a DFGVariable from a DFG using its underlying integer ID.

source

getVariable(dfg, label)

Get a DFGVariable from a DFG using its label.

source

DistributedFactorGraphs.GraphsJl.getFactor — Function.

getFactor(dfg, factorId)

Get a DFGFactor from a DFG using its underlying integer ID.

source

getFactor(dfg, label)

Get a DFGFactor from a DFG using its label.

source

DistributedFactorGraphs.GraphsJl.getNeighbors — Function.

getNeighbors(dfg, node)

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

getNeighbors(dfg, label)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.GraphsJl.ls — Function.

ls(dfg)
ls(dfg, regexFilter)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables.

source

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

source

ls(dfg, label)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

source

DistributedFactorGraphs.GraphsJl.lsf — Function.

lsf(dfg)
lsf(dfg, regexFilter)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

source

Updating Factor Graph Nodes

DistributedFactorGraphs.GraphsJl.updateVariable! — Function.

updateVariable!(dfg, variable)

Update a complete DFGVariable in the DFG.

source

DistributedFactorGraphs.GraphsJl.updateFactor! — Function.

updateFactor!(dfg, factor)

Update a complete DFGFactor in the DFG.

source

Deleting Factor Graph Nodes

DistributedFactorGraphs.GraphsJl.deleteVariable! — Function.

deleteVariable!(dfg, label)

Delete a DFGVariable from the DFG using its label.

source

deleteVariable!(dfg, variable)

Delete a referenced DFGVariable from the DFG.

source

DistributedFactorGraphs.GraphsJl.deleteFactor! — Function.

deleteFactor!(dfg, label)

Delete a DFGFactor from the DFG using its label.

source

deleteFactor!(dfg, factor)

Delete the referened DFGFactor from the DFG.

source

Getting Adjacency Matrix

DistributedFactorGraphs.GraphsJl.getAdjacencyMatrix — Function.

getAdjacencyMatrix(dfg)

Get an adjacency matrix for the DFG, returned as a Matrix{Union{Nothing, Symbol}}. Rows are all factors, columns are all variables, and each cell contains either nothing or the symbol of the relating factor. The first row and first column are factor and variable headings respectively.

source

Validating Factor Graph Connectivity

DistributedFactorGraphs.GraphsJl.isFullyConnected — Function.

isFullyConnected(dfg)

Checks if the graph is fully connected, returns true if so.

source

DistributedFactorGraphs.GraphsJl.hasOrphans — Function.

hasOrphans(dfg)

Checks if the graph is not fully connected, returns true if it is not contiguous.

source

Copying Subgraphs

DistributedFactorGraphs.GraphsJl.getSubgraphAroundNode — Function.

getSubgraphAroundNode(dfg, node)
getSubgraphAroundNode(dfg, node, distance)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors)
getSubgraphAroundNode(dfg, node, distance, includeOrphanFactors, addToDFG)

Retrieve a deep subgraph copy around a given variable or factor. Optionally provide a distance to specify the number of edges should be followed. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

DistributedFactorGraphs.GraphsJl.getSubgraph — Function.

getSubgraph(dfg, variableFactorLabels)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors)
getSubgraph(dfg, variableFactorLabels, includeOrphanFactors, addToDFG)

Get a deep subgraph copy from the DFG given a list of variables and factors. Optionally provide an existing subgraph addToDFG, the extracted nodes will be copied into this graph. By default a new subgraph will be created. Note: By default orphaned factors (where the subgraph does not contain all the related variables) are not returned. Set includeOrphanFactors to return the orphans irrespective of whether the subgraph contains all the variables.

source

DataFrame Extension Functions

Missing docstring.

Missing docstring for getAdjacencyMatrixDataFrame. Check Documenter's build log for details.