DistributedFactorGraphs Functions

DistributedFactorGraphs Functions

Abstract DFG

Common Accessors

DFG Variable Accessors CRUD and SET opperations

DFG Factor Accessors CRUD and SET opperations

Printing

Compare Utilities

Common Functions

Serialization

Abstract DFG

DistributedFactorGraphs._getDuplicatedEmptyDFG — Method

_getDuplicatedEmptyDFG(dfg)

Gets an empty and unique DFG derived from an existing DFG.

DistributedFactorGraphs.addFactor! — Method

addFactor!(dfg, _)

Add a FactorCompute to a DFG.

DistributedFactorGraphs.addFactors! — Method

addFactors!(dfg, factors)

Add a Vector{FactorCompute} to a DFG.

DistributedFactorGraphs.addVariable! — Method

addVariable!(dfg, _)

Add a VariableCompute to a DFG.

DistributedFactorGraphs.addVariables! — Method

addVariables!(dfg, variables)

Add a Vector{VariableCompute} to a DFG.

DistributedFactorGraphs.buildSubgraph — Method

buildSubgraph(, dfg, variableFactorLabels; ...)
buildSubgraph(
    ,
    dfg,
    variableFactorLabels,
    distance;
    solvable,
    graphLabel,
    sessionId,
    kwargs...
)

Build a deep subgraph copy from the DFG given a list of variables and factors and an optional distance. Note: Orphaned factors (where the subgraph does not contain all the related variables) are not returned. Related:

copyGraph!

listNeighborhood

deepcopyGraph

mergeGraph!

Dev Notes

Bulk vs node for node: a list of labels are compiled and the sugraph is copied in bulk.

DistributedFactorGraphs.copyGraph! — Function

copyGraph!(destDFG, sourceDFG; ...)
copyGraph!(destDFG, sourceDFG, variableLabels; ...)
copyGraph!(
    destDFG,
    sourceDFG,
    variableLabels,
    factorLabels;
    copyGraphMetadata,
    overwriteDest,
    deepcopyNodes,
    verbose,
    showprogress
)

Common function for copying nodes from one graph into another graph. This is overridden in specialized implementations for performance. Orphaned factors are not added, with a warning if verbose. Set overwriteDest to overwrite existing variables and factors in the destination DFG. NOTE: copyGraphMetadata not supported yet. Related:

deepcopyGraph

deepcopyGraph!

buildSubgraph

listNeighborhood

mergeGraph!

DistributedFactorGraphs.deepcopyGraph! — Function

deepcopyGraph!(destDFG, sourceDFG; ...)
deepcopyGraph!(destDFG, sourceDFG, variableLabels; ...)
deepcopyGraph!(
    destDFG,
    sourceDFG,
    variableLabels,
    factorLabels;
    kwargs...
)

Copy nodes from one graph into another graph by making deepcopies. see copyGraph! for more detail. Related:

deepcopyGraph

buildSubgraph

listNeighborhood

mergeGraph!

DistributedFactorGraphs.deepcopyGraph — Method

deepcopyGraph(, sourceDFG; ...)
deepcopyGraph(, sourceDFG, variableLabels; ...)
deepcopyGraph(
    ,
    sourceDFG,
    variableLabels,
    factorLabels;
    graphLabel,
    sessionId,
    kwargs...
)

Copy nodes from one graph into a new graph by making deepcopies. see copyGraph! for more detail. Related:

deepcopyGraph!

buildSubgraph

listNeighborhood

mergeGraph!

DistributedFactorGraphs.deleteFactor! — Method

deleteFactor!(dfg, label)

Delete a FactorCompute from the DFG using its label.

DistributedFactorGraphs.deleteFactor! — Method

deleteFactor!(dfg, factor; suppressGetFactor)

Delete the referened FactorCompute from the DFG.

DistributedFactorGraphs.deleteVariable! — Method

deleteVariable!(dfg, variable)

Delete a referenced VariableCompute from the DFG.

Notes

Returns Tuple{AbstractDFGVariable, Vector{<:AbstractDFGFactor}}

DistributedFactorGraphs.deleteVariable! — Method

deleteVariable!(dfg, label)

Delete a VariableCompute from the DFG using its label.

DistributedFactorGraphs.exists — Method

exists(dfg, label)

True if a variable or factor with label exists in the graph.

DistributedFactorGraphs.findClosestTimestamp — Method

findClosestTimestamp(setA, setB)

Find and return the closest timestamp from two sets of Tuples. Also return the minimum delta-time (::Millisecond) and how many elements match from the two sets are separated by the minimum delta-time.

DistributedFactorGraphs.findFactorsBetweenNaive — Function

findFactorsBetweenNaive(dfg, from, to)
findFactorsBetweenNaive(dfg, from, to, assertSingles)

Relatively naive function counting linearly from-to

DevNotes

Convert to using Graphs shortest path methods instead.

DistributedFactorGraphs.findShortestPathDijkstra — Function

Speciallized function available to only GraphsDFG at this time.

Notes

Has option for various types of filters (increases memory usage)

Example

using IncrementalInference

# canonical example graph as example
fg = generateGraph_Kaess()

@show path = findShortestPathDijkstra(fg, :x1, :x3)
@show isVariable.(fg, path)
@show isFactor.(fg, path)

DevNotes

TODO expand to other AbstractDFG entities.

TODO use of filter resource consumption can be improved.

Related

findFactorsBetweenNaive, Graphs.dijkstra_shortest_paths

DistributedFactorGraphs.findVariableNearTimestamp — Function

findVariableNearTimestamp(dfg, timest; ...)
findVariableNearTimestamp(
    dfg,
    timest,
    regexFilter;
    tags,
    solvable,
    warnDuplicate,
    number
)

Find and return nearest variable labels per delta time. Function will filter on regexFilter, tags, and solvable.

Notes

Returns Vector{Tuple{Vector{Symbol}, Millisecond}}

DevNotes:

TODO number should allow returning more than one for k-nearest matches.

Future versions likely will require some optimization around the internal getVariable call.

Perhaps a dedicated/efficient getVariableTimestamp for all DFG flavors.

Related

ls, listVariables, findClosestTimestamp

DistributedFactorGraphs.getAddHistory — Method

getAddHistory(dfg)

DistributedFactorGraphs.getAdjacencyMatrixSymbols — Method

getAdjacencyMatrixSymbols(dfg; solvable)

Get a matrix indicating relationships between variables and factors. Rows are all factors, columns are all variables, and each cell contains either nothing or the symbol of the relating factor. The first row and first column are factor and variable headings respectively. Note:

rather use getBiadjacencyMatrix

Returns either of ::Matrix{Union{Nothing, Symbol}}

DistributedFactorGraphs.getAgent — Method

getAgent(dfg)

DistributedFactorGraphs.getAgentLabel — Method

getAgentLabel(dfg)

DistributedFactorGraphs.getAgentMetadata — Method

getAgentMetadata(dfg)

Get the metadata from the agent in the AbstractDFG.

DistributedFactorGraphs.getBiadjacencyMatrix — Method

getBiadjacencyMatrix(dfg; solvable)

Get a matrix indicating adjacency between variables and factors. Returned as a named tuple: B::SparseMatrixCSC{Int}, varLabels::Vector{Symbol) facLabels::Vector{Symbol). Rows are the factors, columns are the variables, with the corresponding labels in varLabels,facLabels.

Notes

Returns ::NamedTuple{(:B, :varLabels, :facLabels), Tuple{SparseMatrixCSC, Vector{Symbol}, Vector{Symbol}}}

DistributedFactorGraphs.getDFGInfo — Method

getDFGInfo(dfg)

Convenience function to get all the metadata of a DFG

DistributedFactorGraphs.getDescription — Method

getDescription(dfg)

DistributedFactorGraphs.getFactor — Method

getFactor(dfg, label)

Get a FactorCompute from a DFG using its label.

DistributedFactorGraphs.getFactors — Method

getFactors(dfg; ...)
getFactors(dfg, regexFilter; tags, solvable)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

DistributedFactorGraphs.getFactorsSkeleton — Function

Get the skeleton factors from a DFG as a Vector{FactorSkeleton}.

DistributedFactorGraphs.getGraph — Function

DistributedFactorGraphs.getGraphLabel — Method

getGraphLabel(dfg)

DistributedFactorGraphs.getGraphMetadata — Method

getGraphMetadata(dfg)

Get the metadata from the factorgraph in the AbstractDFG.

DistributedFactorGraphs.getId — Method

getId(node)

Get the id of the node.

DistributedFactorGraphs.getLabel — Method

getLabel(node)

Get the label of the node.

DistributedFactorGraphs.getMetadata — Method

getMetadata(node)

Get the metadata of the node.

DistributedFactorGraphs.getSolverParams — Method

getSolverParams(dfg)

DistributedFactorGraphs.getSummaryGraph — Method

getSummaryGraph(dfg)

Get a summary graph (first-class citizens of variables and factors) with the same structure as the original graph.

Notes

this is a copy of the original.

Returns ::GraphsDFG{NoSolverParams, VariableSummary, FactorSummary}

DistributedFactorGraphs.getTypeDFGFactors — Function

Function to get the type of the factors in the DFG.

DistributedFactorGraphs.getTypeDFGVariables — Function

Function to get the type of the variables in the DFG.

DistributedFactorGraphs.getVariable — Method

getVariable(dfg, label, solveKey)

Get a VariableCompute with a specific solver key. In memory types still return a reference, other types returns a variable with only solveKey.

DistributedFactorGraphs.getVariable — Method

getVariable(dfg, label)

Get a VariableCompute from a DFG using its label.

DistributedFactorGraphs.getVariableSkeleton — Function

Get a VariableSkeleton from a DFG.

DistributedFactorGraphs.getVariableSummary — Function

Get a VariableSummary from a DFG.

DistributedFactorGraphs.getVariables — Method

getVariables(dfg; ...)
getVariables(dfg, regexFilter; tags, solvable)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

DistributedFactorGraphs.getVariablesSkeleton — Function

Get the variables from a DFG as a Vector{VariableSkeleton}.

DistributedFactorGraphs.getVariablesSummary — Function

Get the variables from a DFG as a Vector{VariableSummary}.

DistributedFactorGraphs.hasFactor — Method

hasFactor(dfg, label)

True if the factor exists in the graph.

DistributedFactorGraphs.hasTags — Method

hasTags(dfg, sym, tags; matchAll)

Determine if the variable or factor neighbors have the tags:;Vector{Symbol}, and matchAll::Bool.

DistributedFactorGraphs.hasTagsNeighbors — Method

hasTagsNeighbors(dfg, sym, tags; matchAll)

Determine if the variable or factor neighbors have the tags:;Vector{Symbol}, and matchAll::Bool.

DistributedFactorGraphs.hasVariable — Method

hasVariable(dfg, label)

True if the variable exists in the graph.

DistributedFactorGraphs.isConnected — Method

isConnected(dfg)

Checks if the graph is fully connected, returns true if so.

DistributedFactorGraphs.isFactor — Method

isFactor(dfg, sym)

Return whether sym::Symbol represents a factor vertex in the graph DFG. Checks whether it both exists in the graph and is a factor. (If you rather want a quicker for type, just do node isa FactorCompute)

DistributedFactorGraphs.isPathFactorsHomogeneous — Method

isPathFactorsHomogeneous(dfg, from, to)

Return (::Bool,::Vector{TypeName}) of types between two nodes in the factor graph

DevNotes

Only works on LigthDFG at the moment.

Related

findShortestPathDijkstra

DistributedFactorGraphs.isVariable — Method

isVariable(dfg, sym)

Return whether sym::Symbol represents a variable vertex in the graph DFG. Checks whether it both exists in the graph and is a variable. (If you rather want a quick for type, just do node isa VariableCompute)

DistributedFactorGraphs.listFactors — Function

listFactors(dfg; ...)
listFactors(dfg, regexFilter; tags, solvable)

Get a list of the labels of the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

DistributedFactorGraphs.listNeighborhood — Method

listNeighborhood(dfg, label, distance)

Build a list of all unique neighbors inside 'distance'

Notes

Returns Vector{Symbol}

Related:

copyGraph!

buildSubgraph

deepcopyGraph

mergeGraph!

DistributedFactorGraphs.listNeighbors — Method

listNeighbors(dfg, label; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label.

DistributedFactorGraphs.listSolveKeys — Function

listSolveKeys(variable::VariableCompute) -> Set{Symbol}
listSolveKeys(
    variable::VariableCompute,
    filterSolveKeys::Union{Nothing, Regex}
) -> Set{Symbol}
listSolveKeys(
    variable::VariableCompute,
    filterSolveKeys::Union{Nothing, Regex},
    skeys
) -> Any

List all the solvekeys used amongst all variables in the distributed factor graph object.

Related

listSolveKeys, getSolverDataDict, listVariables

DistributedFactorGraphs.listVariables — Function

listVariables(dfg; ...)
listVariables(dfg, regexFilter; tags, solvable)

Get a list of labels of the DFGVariables in the graph. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

Notes

Returns ::Vector{Symbol}

Example

listVariables(dfg, r"l", tags=[:APRILTAG;])

See also: ls

DistributedFactorGraphs.ls — Method

ls(dfg; ...)
ls(dfg, regexFilter; tags, solvable)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

Notes:

Returns Vector{Symbol}

DistributedFactorGraphs.ls — Method

ls(dfg; ...)
ls(dfg, node; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

DistributedFactorGraphs.ls2 — Method

ls2(dfg, label)

Helper to return neighbors at distance 2 around a given node.

DistributedFactorGraphs.lsWho — Method

lsWho(dfg, type)

Gives back all factor labels that fit the bill: lsWho(dfg, :Pose3)

Notes

Returns Vector{Symbol}

Dev Notes

Cloud versions will benefit from less data transfer

ls(dfg::C, ::T) where {C <: CloudDFG, T <: ..}

Related

ls, lsf, lsfPriors

DistributedFactorGraphs.lsf — Method

lsf(dfg; ...)
lsf(dfg, regexFilter; tags, solvable)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

Notes

Return Vector{Symbol}

DistributedFactorGraphs.lsf — Method

lsf(dfg; ...)
lsf(dfg, _)

Lists the factors of a specific type in the factor graph. Example, list all the Point2Point2 factors in the factor graph dfg: lsf(dfg, Point2Point2)

Notes

Return Vector{Symbol}

DistributedFactorGraphs.lsfPriors — Method

lsfPriors(dfg)

Return vector of prior factor symbol labels in factor graph dfg.

Notes:

Returns Vector{Symbol}

DistributedFactorGraphs.mergeFactor! — Method

mergeFactor!(dfg, factor)

Merge a factor into the DFG. If a factor with the same label exists, it will be overwritten; otherwise, the factor will be added to the graph.

DistributedFactorGraphs.mergeGraph! — Function

mergeGraph!(destDFG, sourceDFG; ...)
mergeGraph!(destDFG, sourceDFG, variableLabels; ...)
mergeGraph!(
    destDFG,
    sourceDFG,
    variableLabels,
    factorLabels;
    ...
)
mergeGraph!(
    destDFG,
    sourceDFG,
    variableLabels,
    factorLabels,
    distance;
    solvable,
    kwargs...
)

Merger sourceDFG to destDFG given an optional list of variables and factors and distance. Notes:

Nodes already in the destination graph are updated from sourceDFG.

Orphaned factors (where the subgraph does not contain all the related variables) are not included.

Related:

copyGraph!

buildSubgraph

listNeighborhood

deepcopyGraph

DistributedFactorGraphs.mergeVariable! — Method

mergeVariable!(dfg, variable)

Merge a variable into the DFG. If a variable with the same label exists, it will be overwritten; otherwise, the variable will be added to the graph.

DistributedFactorGraphs.rebuildFactorCache! — Function

rebuildFactorCache!(dfg, factor)
rebuildFactorCache!(dfg, factor, neighbors)

Method must be overloaded by the user for Serialization to work.

DistributedFactorGraphs.setAgentMetadata! — Method

setAgentMetadata!(dfg, data)

Set the metadata of the agent in the AbstractDFG.

DistributedFactorGraphs.setDescription! — Method

setDescription!(dfg, description)

DistributedFactorGraphs.setGraphMetadata! — Method

setGraphMetadata!(dfg, data)

Set the metadata of the factorgraph in the AbstractDFG.

DistributedFactorGraphs.setMetadata! — Method

setMetadata!(node, metadata)

Set the metadata of the node.

DistributedFactorGraphs.toDot — Method

toDot(dfg)

Produces a dot-format of the graph for visualization.

Notes

Returns ::String

DistributedFactorGraphs.toDotFile — Function

toDotFile(dfg)
toDotFile(dfg, fileName)

Produces a dot file of the graph for visualization. Download XDot to see the data

Note

Default location "/tmp/dfg.dot" – MIGHT BE REMOVED

Can be viewed with the xdot system application.

Based on graphviz.org

Common Accessors

Common Accessors to both variable and factor nodes

DistributedFactorGraphs.emptyTags! — Method

emptyTags!(dfg, sym)

Empty all tags from the node (empty)

DistributedFactorGraphs.getLabel — Method

getLabel(v)

Return the label for a DFGNode.

DistributedFactorGraphs.getSolvable — Method

getSolvable(dfg, sym)

Get 'solvable' parameter for either a variable or factor.

DistributedFactorGraphs.getSolvable — Method

getSolvable(var)

Variables or factors may or may not be 'solvable', depending on a user definition. Useful for ensuring atomic transactions.

Related:

isSolveInProgress

DistributedFactorGraphs.getSolveInProgress — Function

getSolveInProgress(var)
getSolveInProgress(var, solveKey)

Which variables or factors are currently being used by an active solver. Useful for ensuring atomic transactions.

DevNotes:

Will be renamed to data.solveinprogress which will be in VND, not DFGNode – see DFG #201

Related

isSolvable

DistributedFactorGraphs.getTags — Method

getTags(v)

Return the tags for a DFGNode.

DistributedFactorGraphs.getTimestamp — Method

getTimestamp(v)

Get the timestamp of a DFGNode.

DistributedFactorGraphs.isSolvable — Method

isSolvable(node)

Variables or factors may or may not be 'solvable', depending on a user definition. returns true if getSolvable > 0 Related:

getSolvable(@ref)

DistributedFactorGraphs.listTags — Method

listTags(dfg, sym)

Return the tags for a variable or factor.

DistributedFactorGraphs.mergeTags! — Method

mergeTags!(dfg, sym, tags)

Merge add tags to a variable or factor (union)

DistributedFactorGraphs.removeTags! — Method

removeTags!(dfg, sym, tags)

Remove the tags from the node (setdiff)

DistributedFactorGraphs.setSolvable! — Method

setSolvable!(dfg, sym, solvable)

Set the solvable parameter for either a variable or factor.

DistributedFactorGraphs.setSolvable! — Method

setSolvable!(node, solvable)

Set the solvable parameter for either a variable or factor.

DistributedFactorGraphs.setTags! — Method

setTags!(f, tags)

Set the tags for a DFGNode.

DistributedFactorGraphs.setTimestamp! — Method

setTimestamp!(dfg, lbl, ts)

Set the timestamp of a Variable/Factor object in a factor graph. Note: Since timestamp is not mutable setTimestamp! calls mergeVariable! internally. See also setTimestamp

DFG Variable Accessors CRUD and SET opperations

DistributedFactorGraphs.addMetadata! — Method

addMetadata!(dfg, label, pair)

Add a Metadata pair key=>value for variable label in dfg

DistributedFactorGraphs.addPPE! — Function

addPPE!(dfg, sourceVariable)
addPPE!(dfg, sourceVariable, ppekey)

Add a new PPE entry from a deepcopy of the source variable PPE. NOTE: Copies the PPE.

DistributedFactorGraphs.addPPE! — Method

addPPE!(dfg, variablekey, ppe)

Add variable PPE, errors if it already exists.

DistributedFactorGraphs.addVariableState! — Method

addVariableState!(dfg, variablekey, state)

Add variable solver data, errors if it already exists.

DistributedFactorGraphs.deleteMetadata! — Method

deleteMetadata!(dfg, label, key)

Delete a Metadata entry at key for variable label in dfg

DistributedFactorGraphs.deletePPE! — Function

deletePPE!(dfg, variablekey)
deletePPE!(dfg, variablekey, ppekey)

Delete PPE data, returns the deleted element.

DistributedFactorGraphs.deletePPE! — Function

deletePPE!(dfg, sourceVariable)
deletePPE!(dfg, sourceVariable, ppekey)

Delete PPE data, returns the deleted element.

DistributedFactorGraphs.deleteVariableState! — Method

deleteVariableState!(dfg, variablekey, solveKey)

Delete variable solver data, returns the number of deleted elements.

DistributedFactorGraphs.deleteVariableState! — Method

deleteVariableState!(dfg, sourceVariable, solveKey)

Delete variable solver data, returns the number of deleted elements.

DistributedFactorGraphs.emptyMetadata! — Method

emptyMetadata!(dfg, label)

Empty all Metadata from variable label in dfg

DistributedFactorGraphs.getCoordinates — Method

getCoordinates(, p)
getCoordinates(, p, basis)

Default reduction of a variable point value (a group element) into coordinates as Vector. Override if defaults are not correct.

DevNotes

TODO Likely remove as part of serialization updates, see #590

Used in transition period for Serialization. This function will likely be changed or deprecated entirely.

Related

getPoint

DistributedFactorGraphs.getDimension — Function

Interface function to return the variableType dimension of an VariableStateType, extend this function for all Types<:VariableStateType.

DistributedFactorGraphs.getLastUpdatedTimestamp — Method

getLastUpdatedTimestamp(est)

DistributedFactorGraphs.getManifold — Method

getManifold(_)

Interface function to return the <:ManifoldsBase.AbstractManifold object of variableType<:VariableStateType.

DistributedFactorGraphs.getMetadata — Method

getMetadata(dfg, label, key)

Get the Metadata entry at key for variable label in dfg

DistributedFactorGraphs.getPPE — Function

getPPE(vari)
getPPE(vari, solveKey)

Get the parametric point estimate (PPE) for a variable in the factor graph.

Notes

Defaults on keywords solveKey and method

Related

getMeanPPE, getMaxPPE, getKDEMean, getKDEFit, getPPEs, getVariablePPEs

DistributedFactorGraphs.getPPE — Function

getPPE(v)
getPPE(v, ppekey)

Get the parametric point estimate (PPE) for a variable in the factor graph for a given solve key.

Notes

Defaults on keywords solveKey and method

Related getPPEMean, getPPEMax, updatePPE!, mean(BeliefType)

DistributedFactorGraphs.getPPEDict — Method

getPPEDict(v)

Get the PPE dictionary for a variable. Recommended to use CRUD operations instead, getPPE, addPPE!, updatePPE!, deletePPE!.

DistributedFactorGraphs.getPPEMax — Method

getPPEMax(est)

DistributedFactorGraphs.getPPEMean — Method

getPPEMean(est)

DistributedFactorGraphs.getPPESuggested — Method

getPPESuggested(est)

DistributedFactorGraphs.getPPEs — Function

Get all the parametric point estimate (PPE) for a variable in the factor graph.

DistributedFactorGraphs.getPoint — Method

getPoint(, v)
getPoint(, v, basis)

Default escalzation from coordinates to a group representation point. Override if defaults are not correct. E.g. coords -> se(2) -> SE(2).

DevNotes

TODO Likely remove as part of serialization updates, see #590

Used in transition period for Serialization. This function will likely be changed or deprecated entirely.

Related

getCoordinates

DistributedFactorGraphs.getPointIdentity — Function

Interface function to return the user provided identity point for this VariableStateType manifold, extend this function for all Types<:VariableStateType.

Notes

Used in transition period for Serialization. This function will likely be changed or deprecated entirely.

DistributedFactorGraphs.getPointType — Function

Interface function to return the manifold point type of an VariableStateType, extend this function for all Types<:VariableStateType.

DistributedFactorGraphs.getSolvedCount — Method

getSolvedCount(v)

Get the number of times a variable has been inferred – i.e. solvedCount.

Related

isSolved, setSolvedCount!

DistributedFactorGraphs.getSolverDataDict — Method

getSolverDataDict(v)

Get solver data dictionary for a variable. Advised to use graph CRUD operations instead.

DistributedFactorGraphs.getVariablePPE — Method

getVariablePPE(::VariableCompute)
getVariablePPE(::VariableState)

Get the Parametric Point Estimate of the given variable.

DistributedFactorGraphs.getVariablePPEDict — Method

getVariablePPEDict(vari)

Return full dictionary of PPEs in a variable, recommended to rather use CRUD: getPPE,

DistributedFactorGraphs.getVariableState — Method

getVariableState(dfg, variableLabel, label)

Get variable solverdata for a given solve key.

DistributedFactorGraphs.getVariableState — Method

getVariableState(v, label)

Retrieve solver data structure stored in a variable.

DistributedFactorGraphs.getVariableType — Method

getVariableType(_)

Variable nodes variableType information holding a variety of meta data associated with the type of variable stored in that node of the factor graph.

Notes

API Quirk in that this function returns and instance of ::T not a ::Type{<:VariableStateType}.

DevWork

TODO, see IncrementalInference.jl 1228

Related

getVariableType

DistributedFactorGraphs.getVariableTypeName — Method

getVariableTypeName(v)

Retrieve the soft type name symbol for a VariableSummary. ie :Point2, Pose2, etc.

DistributedFactorGraphs.isInitialized — Function

isInitialized(var)
isInitialized(var, key)

Returns state of variable data .initialized flag.

Notes:

used by both factor graph variable and Bayes tree clique logic.

DistributedFactorGraphs.isMarginalized — Function

isMarginalized(vert)
isMarginalized(vert, solveKey)

Return ::Bool on whether this variable has been marginalized.

Notes:

VariableState default solveKey=:default

DistributedFactorGraphs.isSolved — Method

isSolved(v)

Boolean on whether the variable has been solved.

Related

getSolved, setSolved!

DistributedFactorGraphs.listMetadata — Method

listMetadata(dfg, label)

List all Metadata keys for a variable label in dfg

DistributedFactorGraphs.listPPEs — Method

listPPEs(dfg, variablekey)

List all the PPE data keys in the variable.

DistributedFactorGraphs.listVariableStates — Method

listVariableStates(dfg, variablekey)

List all the solver data keys in the variable.

DistributedFactorGraphs.mergePPEs! — Method

mergePPEs!(destVariable, sourceVariable)

Merges and updates solver and estimate data for a variable (variable can be from another graph). Note: Makes a copy of the estimates and solver data so that there is no coupling between graphs.

DistributedFactorGraphs.mergeVariableState! — Method

mergeVariableState!(dfg, variablekey, vnd)

Update the variable state if it exists, otherwise add it.

Related

mergeVariableStates!

DistributedFactorGraphs.setMarginalized! — Method

setMarginalized!(vnd, val)

Mark a variable as marginalized true or false.

DistributedFactorGraphs.setSolvedCount! — Method

setSolvedCount!(v, val)

Update/set the solveCount value.

Related

getSolved, isSolved

DistributedFactorGraphs.setTimestamp — Method

setTimestamp(v, ts; verbose)

Set the timestamp of a VariableCompute object returning a new VariableCompute. Note: Since the timestamp field is not mutable setTimestamp returns a new variable with the updated timestamp (note the absence of !). Use mergeVariable! on the returened variable to update it in the factor graph if needed. Alternatively use setTimestamp!. See issue #315.

DistributedFactorGraphs.updateMetadata! — Method

updateMetadata!(dfg, label, pair; warn_if_absent)

Update a Metadata pair key=>value for variable label in dfg

DistributedFactorGraphs.updatePPE! — Function

updatePPE!(dfg, sourceVariable; ...)
updatePPE!(dfg, sourceVariable, ppekey; warn_if_absent)

Update PPE data if it exists, otherwise add it. NOTE: Copies the PPE data.

DistributedFactorGraphs.updatePPE! — Function

updatePPE!(dfg, sourceVariables; ...)
updatePPE!(dfg, sourceVariables, ppekey; warn_if_absent)

Update PPE data if it exists, otherwise add it.

DistributedFactorGraphs.updatePPE! — Method

updatePPE!(dfg, variablekey, ppe; warn_if_absent)

Update PPE data if it exists, otherwise add it – one call per key::Symbol=:default.

Notes

uses ppe.solveKey as solveKey.

DistributedFactorGraphs.@defVarstateType — Macro

@defVarstateType StructName manifold point_identity

A macro to create a new variable type with name StructName associated with a given manifold and identity point.

StructName is the name of the new variable type, which will be defined as a subtype of VariableStateType.

manifold is an object that must be a subtype of ManifoldsBase.AbstractManifold.

point_identity is the identity point on the manifold, used as a reference for operations.

This macro is useful for defining variable types that are not parameterized by dimension, and for associating them with a specific manifold and identity point.

See the Manifolds.jl documentation on creating your own manifolds for more information.

Example:

DFG.@defVariable Pose2 SpecialEuclidean(2) ArrayPartition([0;0.0],[1 0; 0 1.0])

DistributedFactorGraphs.@defVarstateTypeN — Macro

@defVarstateTypeN StructName manifold point_identity

A macro to create a new variable type with name StructName that is parameterized by N and associated with a given manifold and identity point.

StructName is the name of the new variable type, which will be defined as a subtype of VariableStateType{N}.

manifold is an object that must be a subtype of ManifoldsBase.AbstractManifold.

point_identity is the identity point on the manifold, used as a reference for operations.

This macro is useful for defining variable types that are parameterized by dimension or other type parameters (e.g., Pose{N}), and for associating them with a specific manifold and identity point.

See the Manifolds.jl documentation on creating your own manifolds for more information.

Example:

DFG.@defVarstateTypeN Pose{N} SpecialEuclidean(N) ArrayPartition(zeros(SVector{N, Float64}), SMatrix{N, N, Float64}(I))

DFG Factor Accessors CRUD and SET opperations

DistributedFactorGraphs._getPriorType — Method

_getPriorType(_type)

If you know a variable is ::Type{<:Pose2} but want to find its default prior ::Type{<:PriorPose2}.

Assumptions

The prior type will be defined in the same module as the variable type.

Not exported per default, but can be used with knowledge of the caveats.

Example

using RoME
@assert RoME.PriorPose2 == DFG._getPriorType(Pose2)

DistributedFactorGraphs.getCache — Method

getCache(f)

Return the solver cache for a factor, which is used to store intermediate results during the solving process. This is useful for caching results that can be reused across multiple solves, such as Jacobians or other computed values.

DistributedFactorGraphs.getFactorFunction — Method

getFactorFunction(fc)

Return reference to the user factor in <:AbstractDFG identified by ::Symbol.

DistributedFactorGraphs.getFactorState — Method

getFactorState(f)

Return factor state from factor graph.

DistributedFactorGraphs.getFactorType — Method

getFactorType(fct)

Return user factor type from factor graph identified by label ::Symbol.

Notes

Replaces older getfnctype.

DistributedFactorGraphs.getObservation — Method

getObservation(f)

Return the observation of a factor, which is the user-defined data structure that contains the information about the factor, such as the measurement, prior, or relative pose.

DistributedFactorGraphs.getVariableOrder — Method

getVariableOrder(fct)

Get the variable ordering for this factor. Should be equivalent to listNeighbors unless something was deleted in the graph.

DistributedFactorGraphs.isPrior — Method

isPrior(dfg, fc)

Return ::Bool on whether given factor fc::Symbol is a prior in factor graph dfg.

DistributedFactorGraphs.setCache! — Method

setCache!(f, solvercache)

Set the solver cache for a factor, which is used to store intermediate results during the solving process. This is useful for caching results that can be reused across multiple solves, such as Jacobians or other computed values.

DistributedFactorGraphs.@defObservationType — Macro

@defObservationType StructName factortype<:AbstractFactorObservation manifolds<:AbstractManifold

A macro to create a new factor function with name StructName and manifold. Note that the manifold is an object and must be a subtype of ManifoldsBase.AbstractManifold. See documentation in Manifolds.jl on making your own.

Example:

DFG.@defObservationType Pose2Pose2 RelativeObservation SpecialEuclidean(2)

Printing

DistributedFactorGraphs.printFactor — Method

printFactor(dfg, sym; kwargs...)

Display and return to console the user factor identified by tag name.

DistributedFactorGraphs.printVariable — Method

printVariable(dfg, sym; kwargs...)

Display the content of VariableState to console for a given factor graph and variable tag::Symbol.

Dev Notes

TODO split as two show macros between AMP and DFG

Compare Utilities

DistributedFactorGraphs.compareAll — Method

compareAll(Al, Bl; show, skip)

Recursively compare the all fields of T that are not in skip for objects Al and Bl.

TODO > add to func_ref.md

DistributedFactorGraphs.compareAllVariables — Method

compareAllVariables(fgA, fgB; skip, show, skipsamples)

Compare all variables in both ::FactorGraphs A and B.

Notes

A and B should all the same variables and factors.

Related:

compareFactorGraphs, compareSimilarVariables, compareVariable, ls

DistributedFactorGraphs.compareFactor — Method

compareFactor(A, B; show, skip, skipsamples, skipcompute)

Compare that all fields are the same in a ::FactorGraph factor.

DevNotes

TODO getSolverData(A).fnc.varValsAll / varidx are only defined downstream, so should should this function not be in IIF?

DistributedFactorGraphs.compareFactorGraphs — Method

compareFactorGraphs(
    fgA,
    fgB;
    skipsamples,
    skipcompute,
    skip,
    show
)

Compare and return if two factor graph objects are the same, by comparing similar variables and factors.

Notes:

Default items to skip with skipsamples, skipcompute.

User defined fields to skip can be specified with skip::Vector{Symbol}.

To enable debug messages for viewing which fields are not the same:

https://stackoverflow.com/questions/53548681/how-to-enable-debugging-messages-in-juno-julia-editor

Related:

compareSimilarVariables, compareSimilarFactors, compareAllVariables, ls.

DistributedFactorGraphs.compareFields — Method

compareFields(Al, Bl; show, skip)

Compare the all fields of T that are not in skip for objects Al and Bl and returns ::Bool.

TODO > add to func_ref.md

DistributedFactorGraphs.compareSimilarFactors — Method

compareSimilarFactors(
    fgA,
    fgB;
    skipsamples,
    skipcompute,
    skip,
    show
)

Compare similar factors between ::FactorGraphs A and B.

Related:

compareFactorGraphs, compareSimilarVariables, compareAllVariables, ls.

DistributedFactorGraphs.compareSimilarVariables — Method

compareSimilarVariables(fgA, fgB; skip, show, skipsamples)

Compare similar labels between ::FactorGraphs A and B.

Notes

At least one variable label should exist in both A and B.

Related:

compareFactorGraphs, compareAllVariables, compareSimilarFactors, compareVariable, ls.

DistributedFactorGraphs.compareVariable — Method

compareVariable(A, B; skip, show, skipsamples)

Compare that all fields are the same in a ::FactorGraph variable.

Common Functions

DistributedFactorGraphs.getVariableLabelNumber — Function

getVariableLabelNumber(vs)
getVariableLabelNumber(vs, prefix)

Small utility to return ::Int, e.g. 0 from getVariableLabelNumber(:x0)

Examples

getVariableLabelNumber(:l10)          # 10
getVariableLabelNumber(:x1)           # 1
getVariableLabelNumber(:x1_10, "x1_") # 10

DevNotes

make prefix Regex based for longer – i.e. :apriltag578, :lm1_4

DistributedFactorGraphs.isValidLabel — Method

isValidLabel(id)

Returns true if the label is valid for node.

DistributedFactorGraphs.solveGraph! — Function

Default non-parametric graph solution.

DistributedFactorGraphs.solveGraphParametric! — Function

Standard parametric graph solution (Experimental).

DistributedFactorGraphs.sortDFG — Method

sortDFG(vars; by, kwargs...)

Convenience wrapper for Base.sort. Sort variable (factor) lists in a meaningful way (by timestamp, label, etc), for example [:april;:x1_3;:x1_6;] Defaults to sorting by timestamp for variables and factors and using natural_lt for Symbols. See Base.sort for more detail.

Notes

Not fool proof, but does better than native sort.

Example

sortDFG(ls(dfg)) sortDFG(ls(dfg), by=getLabel, lt=natural_lt)

Related

ls, lsf

Serialization

DistributedFactorGraphs.getTypeFromSerializationModule — Method

getTypeFromSerializationModule(_typeString)

Get a type from the serialization module.