DistributedFactorGraphs Functions

DistributedFactorGraphs Functions

Abstract DFG

Common Accessors

Common

DFG Variable Accessors CRUD and SET opperations

DFG Factor Accessors CRUD and SET opperations

Printing

Compare Utilities

Common Functions

Serialization

Abstract DFG

DistributedFactorGraphs.addFactor! — Function

Add a FactorDFG to a DFG. Implement addFactor!(dfg::AbstractDFG, factor::AbstractGraphFactor)

DistributedFactorGraphs.addFactors! — Method

addFactors!(dfg, factors)

Add a Vector{FactorDFG} to a DFG.

DistributedFactorGraphs.addVariable! — Function

Add a VariableDFG to a DFG. Implement addVariable!(dfg::AbstractDFG, variable::AbstractGraphVariable)

DistributedFactorGraphs.addVariables! — Method

addVariables!(dfg, variables)

Add a Vector{VariableDFG} to a DFG. Implement addVariables!(dfg::AbstractDFG, variables::Vector{<:AbstractGraphVariable})

DistributedFactorGraphs.deleteFactor! — Function

Delete a FactorDFG from the DFG using its label. Implement deleteFactor!(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.deleteFactor! — Method

deleteFactor!(dfg, factor)

Delete the referenced Factor from the DFG.

DistributedFactorGraphs.deleteVariable! — Function

Delete a VariableDFG from the DFG. Implement deleteVariable!(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.exists — Method

exists(dfg, label)

True if a variable or factor with label exists in the graph.

DistributedFactorGraphs.findPath — Function

findPath(dfg, from::Symbol, to::Symbol; variableLabels, factorLabels, kwargs...)

Return the single shortest path between from and to. Errors if no path exists (use findPaths for graphs that may be disconnected).

Accepts the same restriction keywords as findPaths.

DistributedFactorGraphs.findPaths — Function

findPaths(dfg, from::Symbol, to::Symbol, k::Int; variableLabels, factorLabels, kwargs...)

Return the k shortest paths between from and to in the factor graph. Each result is a (path = Vector{Symbol}, dist) named tuple.

Optional keyword arguments restrict which variables and/or factors may appear on the path. When neither is given the full graph is used. When only one is provided the other defaults to all labels of that kind in dfg.

Typical usage with filters:

vars = listVariables(dfg; whereSolvable = >=(1))
facs = listFactors(dfg; whereSolvable = >=(1))
findPaths(dfg, :x1, :x5, 3; variableLabels = vars, factorLabels = facs)

See also: findPath, listVariables, listFactors

DistributedFactorGraphs.getAgent — Function

DistributedFactorGraphs.getAgentLabel — Method

getAgentLabel(dfg)

DistributedFactorGraphs.getBiadjacencyMatrix — Method

getBiadjacencyMatrix(dfg; solvable)

Get a matrix indicating adjacency between variables and factors. Returned as a named tuple: B::SparseMatrixCSC{Int}, varLabels::Vector{Symbol) facLabels::Vector{Symbol). Rows are the factors, columns are the variables, with the corresponding labels in varLabels,facLabels.

Notes

Returns ::NamedTuple{(:B, :varLabels, :facLabels), Tuple{SparseMatrixCSC, Vector{Symbol}, Vector{Symbol}}}

DistributedFactorGraphs.getFactor — Function

Get a FactorDFG from a DFG using its label. Implement getFactor(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.getFactors — Function

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

DistributedFactorGraphs.getFactorsSkeleton — Function

Get the skeleton factors from a DFG as a Vector{FactorSkeleton}.

DistributedFactorGraphs.getGraph — Function

DistributedFactorGraphs.getGraphLabel — Method

getGraphLabel(dfg)

DistributedFactorGraphs.getId — Function

DistributedFactorGraphs.getSolverParams — Method

getSolverParams(dfg)

DistributedFactorGraphs.getSubgraph — Method

getSubgraph(, dfg, variableFactorLabels; ...)
getSubgraph(
    ,
    dfg,
    variableFactorLabels,
    distance;
    whereSolvable,
    whereTags,
    graphLabel,
    solvable,
    kwargs...
)

Build a deep subgraph copy from the DFG given a list of variables and factors and an optional distance. Note: Orphaned factors (where the subgraph does not contain all the related variables) are not returned. Related:

listNeighborhood

mergeGraph!

Dev Notes

Bulk vs node for node: a list of labels are compiled and the sugraph is copied in bulk.

DistributedFactorGraphs.getSummaryGraph — Method

getSummaryGraph(dfg)

Get a summary graph (first-class citizens of variables and factors) with the same structure as the original graph.

Notes

this is a copy of the original.

Returns ::GraphsDFG{NoSolverParams, VariableSummary, FactorSummary}

DistributedFactorGraphs.getTypeDFGFactors — Method

getTypeDFGFactors(_)

Function to get the type of the factors in the DFG.

DistributedFactorGraphs.getTypeDFGVariables — Method

getTypeDFGVariables(_)

Function to get the type of the variables in the DFG.

DistributedFactorGraphs.getVariable — Function

Get a VariableDFG from a DFG using its label. Implement getVariable(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.getVariableSkeleton — Function

Get a VariableSkeleton from a DFG.

DistributedFactorGraphs.getVariableSummary — Function

Get a VariableSummary from a DFG.

DistributedFactorGraphs.getVariables — Function

Get the variables in the DFG as a Vector, supporting various filters.

Keyword arguments

whereSolvable: Optional function to filter on the solvable property, eg >=(1).

whereLabel: Optional function to filter on label e.g., contains(r"x1").

whereTags: Optional function to filter on tags, eg. ⊇([:POSE]).

whereType: Optional function to filter on the variable type.

Returns

Vector{<:AbstractGraphVariable} matching the filters.

See also: listVariables, ls

DistributedFactorGraphs.getVariablesSkeleton — Function

Get the variables from a DFG as a Vector{VariableSkeleton}.

DistributedFactorGraphs.getVariablesSummary — Function

Get the variables from a DFG as a Vector{VariableSummary}.

DistributedFactorGraphs.hasFactor — Function

True if the factor exists in the graph. Implement hasFactor(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.hasVariable — Function

True if the variable exists in the graph. Implement hasVariable(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.isConnected — Function

Checks if the graph is fully connected, returns true if so. Implement isConnected(dfg::AbstractDFG)

DistributedFactorGraphs.isFactor — Function

Return whether sym::Symbol represents a factor vertex in the graph DFG. Checks whether it both exists in the graph and is a factor. (If you rather want a quicker for type, just do node isa FactorDFG) Implement isFactor(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.isVariable — Function

Return whether sym::Symbol represents a variable vertex in the graph DFG. Checks whether it both exists in the graph and is a variable. (If you rather want a quick for type, just do node isa VariableDFG) Implement isVariable(dfg::AbstractDFG, label::Symbol)

DistributedFactorGraphs.listNeighborhood — Method

listNeighborhood(dfg, label, distance; filters...)

Build a list of all unique neighbors inside 'distance'. Neighbors can be filtered by using keyword arguments, eg. [whereTags] and [whereSolvable]. Filters are applied to final neighborhood result.

Notes

Returns a tuple (variableLabels, factorLabels), where each element is a Vector{Symbol}.

Related:

getSubgraph

mergeGraph!

DistributedFactorGraphs.listNeighbors — Function

Retrieve a list of labels of the immediate neighbors around a given variable or factor specified by its label. Implement listNeighbors(dfg::AbstractDFG, label::Symbol; whereSolvable, whereTags)

DistributedFactorGraphs.mergeFactor! — Function

Merge a factor into the DFG. If a factor with the same label exists, it will be overwritten; otherwise, the factor will be added to the graph. Implement mergeFactor!(dfg::AbstractDFG, factor::AbstractGraphFactor)

DistributedFactorGraphs.mergeGraph! — Method

mergeGraph!(destDFG, srcDFG)

Merge the source DFG into the destination DFG cascading down the hierarchy of DFG nodes. Merge rules:

Variables, Factors, Agent, and Graphroot, with the same label are merged if they are equal.

On conflicts, a MergeConflictError is thrown.

Child nodes (eg. tags, Bloblets, Blobentries, States, etc.) are using merge!.

The Blobstore links are merged provided they point to the same blobstore.

On conflicts, a MergeConflictError is thrown.

DistributedFactorGraphs.mergeVariable! — Function

Merge a variable into the DFG. If a variable with the same label exists, it will be overwritten; otherwise, the variable will be added to the graph. Implement mergeVariable!(dfg::AbstractDFG, variable::AbstractGraphVariable)

DistributedFactorGraphs.rebuildFactorCache! — Function

rebuildFactorCache!(dfg, factor)
rebuildFactorCache!(dfg, factor, neighbors)

Method must be overloaded by the user for Serialization to work.

DistributedFactorGraphs.toDot — Method

toDot(dfg)

Produces a dot-format of the graph for visualization.

Notes

Returns ::String

DistributedFactorGraphs.toDotFile — Function

toDotFile(dfg)
toDotFile(dfg, fileName)

Produces a dot file of the graph for visualization. Download XDot to see the data

Note

Default location "/tmp/dfg.dot" – MIGHT BE REMOVED

Can be viewed with the xdot system application.

Based on graphviz.org

Common Accessors

Common Accessors to both variable and factor nodes

DistributedFactorGraphs.getDescription — Method

getDescription(node)

DistributedFactorGraphs.getLabel — Method

getLabel(node)

Get the label of the node.

DistributedFactorGraphs.getSolvable — Method

getSolvable(dfg, sym)

Get 'solvable' parameter for either a variable or factor.

DistributedFactorGraphs.getSolvable — Method

getSolvable(node)

Variables or factors may or may not be 'solvable', depending on a user definition. Useful for ensuring atomic transactions.

Related:

isSolveInProgress

DistributedFactorGraphs.getTimestamp — Method

getTimestamp(node)

Get the timestamp of a AbstractGraphNode.

DistributedFactorGraphs.isSolvable — Method

isSolvable(node)

Variables or factors may or may not be 'solvable', depending on a user definition. returns true if getSolvable > 0 Related:

getSolvable(@ref)

DistributedFactorGraphs.setSolvable! — Method

setSolvable!(node, solvable)

Set the solvable parameter for either a variable or factor.

Common

DistributedFactorGraphs.listFactors — Method

listFactors(dfg, args; kwargs...)

Get a list of the labels of the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

DistributedFactorGraphs.listNeighborsSecondary — Method

listNeighborsSecondary(dfg, label)

List the second order neighbors of a given node.

DistributedFactorGraphs.listVariables — Method

listVariables(dfg, args; kwargs...)

Get a list of labels of the DFGVariables in the graph. Supports optional arguments to filter the variables returned.

Notes

Returns ::Vector{Symbol}

Example

listVariables(dfg)

See also: ls

DistributedFactorGraphs.ls — Method

ls(dfg, node; solvable)

Retrieve a list of labels of the immediate neighbors around a given variable or factor.

DistributedFactorGraphs.ls — Method

ls(dfg; whereSolvable, whereTags, whereType, whereLabel)

List the DFGVariables in the DFG. Optionally specify a label regular expression to retrieves a subset of the variables. Tags is a list of any tags that a node must have (at least one match).

Notes:

Returns Vector{Symbol}

DistributedFactorGraphs.lsTypes — Method

lsTypes(dfg)

Return Vector{DataType} of all unique variable types in factor graph.

DistributedFactorGraphs.lsTypesDict — Method

lsTypesDict(dfg)

Return ::Dict{DataType, Vector{Symbol}} of all unique variable types with labels in a factor graph.

DistributedFactorGraphs.lsf — Method

lsf(dfg; whereSolvable, whereTags, whereType, whereLabel)

List the DFGFactors in the DFG. Optionally specify a label regular expression to retrieves a subset of the factors.

Notes

Return Vector{Symbol}

DistributedFactorGraphs.lsf — Method

lsf(dfg, _)

Lists the factors of a specific type in the factor graph. Example, list all the Point2Point2 factors in the factor graph dfg: lsf(dfg, Point2Point2)

Notes

Return Vector{Symbol}

DistributedFactorGraphs.lsfPriors — Method

lsfPriors(dfg)

Return vector of prior factor symbol labels in factor graph dfg.

Notes:

Returns Vector{Symbol}

DistributedFactorGraphs.lsfTypes — Method

lsfTypes(dfg)

Return Vector{Symbol} of all unique factor types in factor graph.

DistributedFactorGraphs.lsfTypesDict — Method

lsfTypesDict(dfg)

Return ::Dict{DataType, Vector{Symbol}} of all unique factors types with labels in a factor graph.

DistributedFactorGraphs.findClosestTimestamp — Method

findClosestTimestamp(setA, setB)

Find and return the closest timestamp from two sets of Tuples. Also return the minimum delta-time (::Nanosecond) and how many elements match from the two sets are separated by the minimum delta-time.

DistributedFactorGraphs.findShortestPathDijkstra — Function

Speciallized function available to only GraphsDFG at this time.

Notes

Has option for various types of filters (increases memory usage)

Example

using IncrementalInference

# canonical example graph as example
fg = generateGraph_Kaess()

@show path = findShortestPathDijkstra(fg, :x1, :x3)
@show isVariable.(fg, path)
@show isFactor.(fg, path)

DevNotes

TODO expand to other AbstractDFG entities.

TODO use of filter resource consumption can be improved.

Related

Graphs.dijkstra_shortest_paths

DistributedFactorGraphs.findVariablesNearTimestamp — Method

findVariablesNearTimestamp(
    dfg,
    query_timestamp;
    whereLabel,
    whereTags,
    whereSolvable,
    number
)

Find and return nearest variable labels per delta time. Function will filter on regexFilter, tags, and solvable.

Notes

Returns Vector{Tuple{Vector{Symbol}, Nanosecond}}

DevNotes:

TODO number should allow returning more than one for k-nearest matches.

Future versions likely will require some optimization around the internal getVariable call.

Perhaps a dedicated/efficient getVariableTimestamp for all DFG flavors.

Related

ls, listVariables, findClosestTimestamp

DistributedFactorGraphs.deleteTags! — Method

deleteTags!(node, tags)

Remove the tags from the node (setdiff)

DistributedFactorGraphs.emptyTags! — Method

emptyTags!(node)

Empty all tags from the node (empty)

DistributedFactorGraphs.hasTags — Method

hasTags(dfg, sym, tags)

Determine if the variable or factor neighbors have the tags::Vector{Symbol}`.

DistributedFactorGraphs.listTags — Method

listTags(node)

DistributedFactorGraphs.listVariableTags — Method

listVariableTags(dfg, sym)

List the tags for a variable.

DistributedFactorGraphs.mergeTags! — Method

mergeTags!(node, tags)

Merge add tags to a variable or factor (union)

DistributedFactorGraphs.addBloblet! — Method

addBloblet!(node, bloblet)

DistributedFactorGraphs.deleteBloblet! — Method

deleteBloblet!(node, label)

DistributedFactorGraphs.getBloblet — Method

getBloblet(node, label)

DistributedFactorGraphs.getBloblets — Method

getBloblets(node)

DistributedFactorGraphs.hasBloblet — Method

hasBloblet(node, label)

Check if a Bloblet with the given label exists on the node.

DistributedFactorGraphs.listBloblets — Method

listBloblets(node)

List all Bloblet keys for a variable label in dfg

DistributedFactorGraphs.mergeBloblet! — Method

mergeBloblet!(node, bloblet)

DistributedFactorGraphs.addAgentBloblet! — Method

addAgentBloblet!(dfg, bloblet)

DistributedFactorGraphs.addAgentBloblets! — Method

addAgentBloblets!(dfg, bloblets)

DistributedFactorGraphs.addFactorBloblet! — Method

addFactorBloblet!(dfg, fac_label, bloblet)

DistributedFactorGraphs.addFactorBloblets! — Method

addFactorBloblets!(dfg, fac_label, bloblets)

DistributedFactorGraphs.addGraphBloblet! — Method

addGraphBloblet!(dfg, bloblet)

DistributedFactorGraphs.addGraphBloblets! — Method

addGraphBloblets!(dfg, bloblets)

DistributedFactorGraphs.addVariableBloblet! — Method

addVariableBloblet!(dfg, var_label, bloblet)

DistributedFactorGraphs.addVariableBloblets! — Method

addVariableBloblets!(dfg, var_label, bloblets)

DistributedFactorGraphs.deleteAgentBloblet! — Method

deleteAgentBloblet!(dfg, label)

DistributedFactorGraphs.deleteAgentBloblets! — Method

deleteAgentBloblets!(dfg, labels)

DistributedFactorGraphs.deleteFactorBloblet! — Method

deleteFactorBloblet!(dfg, fac_label, label)

DistributedFactorGraphs.deleteFactorBloblets! — Method

deleteFactorBloblets!(dfg, fac_label, labels)

DistributedFactorGraphs.deleteGraphBloblet! — Method

deleteGraphBloblet!(dfg, label)

DistributedFactorGraphs.deleteGraphBloblets! — Method

deleteGraphBloblets!(dfg, labels)

DistributedFactorGraphs.deleteVariableBloblet! — Method

deleteVariableBloblet!(dfg, var_label, label)

DistributedFactorGraphs.deleteVariableBloblets! — Method

deleteVariableBloblets!(dfg, var_label, labels)

DistributedFactorGraphs.getAgentBloblet — Method

getAgentBloblet(dfg, label)

DistributedFactorGraphs.getAgentBloblets — Method

getAgentBloblets(dfg)

DistributedFactorGraphs.getFactorBloblet — Method

getFactorBloblet(dfg, fac_label, label)

DistributedFactorGraphs.getFactorBloblets — Method

getFactorBloblets(dfg, fac_label)

DistributedFactorGraphs.getGraphBloblet — Method

getGraphBloblet(dfg, label)

DistributedFactorGraphs.getGraphBloblets — Method

getGraphBloblets(dfg)

DistributedFactorGraphs.getVariableBloblet — Method

getVariableBloblet(dfg, var_label, label)

DistributedFactorGraphs.getVariableBloblets — Method

getVariableBloblets(dfg, var_label)

DistributedFactorGraphs.hasAgentBloblet — Method

hasAgentBloblet(dfg, label)

DistributedFactorGraphs.hasFactorBloblet — Method

hasFactorBloblet(dfg, fac_label, label)

DistributedFactorGraphs.hasGraphBloblet — Method

hasGraphBloblet(dfg, label)

DistributedFactorGraphs.hasVariableBloblet — Method

hasVariableBloblet(dfg, var_label, label)

DistributedFactorGraphs.listAgentBloblets — Method

listAgentBloblets(dfg)

DistributedFactorGraphs.listFactorBloblets — Method

listFactorBloblets(dfg, fac_label)

DistributedFactorGraphs.listGraphBloblets — Method

listGraphBloblets(dfg)

DistributedFactorGraphs.listVariableBloblets — Method

listVariableBloblets(dfg, var_label)

DistributedFactorGraphs.mergeAgentBloblet! — Method

mergeAgentBloblet!(dfg, bloblet)

DistributedFactorGraphs.mergeAgentBloblets! — Method

mergeAgentBloblets!(dfg, bloblets)

DistributedFactorGraphs.mergeFactorBloblet! — Method

mergeFactorBloblet!(dfg, fac_label, bloblet)

DistributedFactorGraphs.mergeFactorBloblets! — Method

mergeFactorBloblets!(dfg, fac_label, bloblets)

DistributedFactorGraphs.mergeGraphBloblet! — Method

mergeGraphBloblet!(dfg, bloblet)

DistributedFactorGraphs.mergeGraphBloblets! — Method

mergeGraphBloblets!(dfg, bloblets)

DistributedFactorGraphs.mergeVariableBloblet! — Method

mergeVariableBloblet!(dfg, var_label, bloblet)

DistributedFactorGraphs.mergeVariableBloblets! — Method

mergeVariableBloblets!(dfg, var_label, bloblets)

DFG Variable Accessors CRUD and SET opperations

DistributedFactorGraphs.addState! — Method

addState!(dfg, variableLabel, state)

Add variable solver data, errors if it already exists.

DistributedFactorGraphs.addStates! — Method

addStates!(dfg, variableLabel, states)

Add variable States by calling addState!. NOTE: If an error occurs while adding one of the states, previously added states will not be rolled back.

DistributedFactorGraphs.deleteState! — Method

deleteState!(dfg, variableLabel, label)

Delete the variable State by label, returns the number of deleted elements.

DistributedFactorGraphs.deleteStates! — Method

deleteStates!(dfg, variableLabel, labels)

Delete the variable States by label, returns the number of deleted elements.

DistributedFactorGraphs.getDimension — Function

Interface function to return the variableType dimension of an StateType, extend this function for all Types<:StateType.

DistributedFactorGraphs.getManifold — Method

getManifold(_)

Interface function to return the <:ManifoldsBase.AbstractManifold object of variableType<:StateType.

DistributedFactorGraphs.getPointIdentity — Function

Interface function to return the user provided identity point for this StateType manifold, extend this function for all Types<:StateType.

Notes

Used in transition period for Serialization. This function will likely be changed or deprecated entirely.

DistributedFactorGraphs.getPointType — Function

Interface function to return the manifold point type of an StateType, extend this function for all Types<:StateType.

DistributedFactorGraphs.getSolvedCount — Method

getSolvedCount(v)

Get the number of times a variable has been inferred – i.e. solvedCount.

Related

isSolved, setSolvedCount!

DistributedFactorGraphs.getState — Method

getState(dfg, variableLabel, label)

Get the variable State for a given state label.

DistributedFactorGraphs.getStateKind — Method

getStateKind(_)

Get the kind of the variable's state, eg. Pose2, Point3, etc. as an instance of StateType.

DistributedFactorGraphs.isInitialized — Function

isInitialized(var)
isInitialized(var, key)

Returns state of variable data .initialized flag.

Notes:

used by both factor graph variable and Bayes tree clique logic.

DistributedFactorGraphs.isMarginalized — Function

isMarginalized(vert)
isMarginalized(vert, solveKey)

Return ::Bool on whether this variable has been marginalized.

Notes:

State default solveKey=:default

DistributedFactorGraphs.isSolved — Method

isSolved(v)

Boolean on whether the variable has been solved.

Related

getSolved, setSolved!

DistributedFactorGraphs.listStates — Method

listStates(v; whereLabel)

List all the variable state labels.

DistributedFactorGraphs.mergeState! — Method

mergeState!(dfg, variableLabel, vnd)

Update the variable state if it exists, otherwise add it.

Related

mergeStates!

DistributedFactorGraphs.setMarginalized! — Method

setMarginalized!(vnd, val)

Mark a variable as marginalized true or false.

DistributedFactorGraphs.setSolvedCount! — Method

setSolvedCount!(v, val)

Update/set the solveCount value.

Related

getSolved, isSolved

DistributedFactorGraphs.@defStateType — Macro

@defStateType StructName manifold point_identity

A macro to create a new variable type with name StructName associated with a given manifold and identity point.

StructName is the name of the new variable type, which will be defined as a subtype of StateType.

manifold is an object that must be a subtype of ManifoldsBase.AbstractManifold.

point_identity is the identity point on the manifold, used as a reference for operations.

This macro is useful for defining variable types that are not parameterized by dimension, and for associating them with a specific manifold and identity point.

See the Manifolds.jl documentation on creating your own manifolds for more information.

Example:

DFG.@defStateType Pose2 SpecialEuclideanGroup(2) ArrayPartition([0;0.0],[1 0; 0 1.0])

DistributedFactorGraphs.@defStateTypeN — Macro

@defStateTypeN StructName manifold point_identity

A macro to create a new variable type with name StructName that is parameterized by N and associated with a given manifold and identity point.

StructName is the name of the new variable type, which will be defined as a subtype of StateType{N}.

manifold is an object that must be a subtype of ManifoldsBase.AbstractManifold.

point_identity is the identity point on the manifold, used as a reference for operations.

This macro is useful for defining variable types that are parameterized by dimension or other type parameters (e.g., Pose{N}), and for associating them with a specific manifold and identity point.

See the Manifolds.jl documentation on creating your own manifolds for more information.

Example:

DFG.@defStateTypeN Pose{N} SpecialEuclideanGroup(N) ArrayPartition(zeros(SVector{N, Float64}), SMatrix{N, N, Float64}(I))

DFG Factor Accessors CRUD and SET opperations

DistributedFactorGraphs._getPriorType — Method

_getPriorType(T)

If you know a variable is ::Type{<:Pose2} but want to find its default prior ::Type{<:PriorPose2}.

Assumptions

The prior type will be defined in the same module as the variable type.

Not exported per default, but can be used with knowledge of the caveats.

Example

using RoME
@assert RoME.PriorPose2 == DFG._getPriorType(Pose2)

DistributedFactorGraphs.getCache — Method

getCache(f)

Return the solver cache for a factor, which is used to store intermediate results during the solving process. This is useful for caching results that can be reused across multiple solves, such as Jacobians or other computed values.

DistributedFactorGraphs.getObservation — Method

getObservation(f)

Return the observation of a factor, which is the user-defined data structure that contains the information about the factor, such as the measurement, prior, or relative pose.

DistributedFactorGraphs.getRecipehyper — Method

getRecipehyper(f)

Return the hyperparameters associated with a factor in the factor graph. These hyperparameters may include settings such as multi-hypothesis handling, null hypothesis thresholds, and inflation factors that influence the behavior of the factor during optimization.

DistributedFactorGraphs.getRecipestate — Method

getRecipestate(f)

Return factor recipe state from factor graph.

DistributedFactorGraphs.getVariableOrder — Method

getVariableOrder(fct)

Get the variable ordering for this factor. Should be equivalent to listNeighbors unless something was deleted in the graph.

DistributedFactorGraphs.isPrior — Method

isPrior(_)
isPrior(_)

Return ::Bool on whether given factor fc::Symbol is a prior in factor graph dfg.

DistributedFactorGraphs.setCache! — Method

setCache!(f, solvercache)

Set the solver cache for a factor, which is used to store intermediate results during the solving process. This is useful for caching results that can be reused across multiple solves, such as Jacobians or other computed values.

DistributedFactorGraphs.@defObservationType — Macro

@defObservationType StructName factortype<:AbstractObservation manifolds<:AbstractManifold

A macro to create a new factor function with name StructName and manifold. Note that the manifold is an object and must be a subtype of ManifoldsBase.AbstractManifold. See documentation in Manifolds.jl on making your own.

Example:

DFG.@defObservationType Pose2Pose2 RelativeObservation SpecialEuclideanGroup(2)

Printing

DistributedFactorGraphs.printFactor — Method

printFactor(dfg, sym; kwargs...)

Display and return to console the user factor identified by tag name.

DistributedFactorGraphs.printVariable — Method

printVariable(dfg, sym; kwargs...)

Display the content of State to console for a given factor graph and variable tag::Symbol.

Dev Notes

TODO split as two show macros between AMP and DFG

Compare Utilities

DistributedFactorGraphs.compareAll — Method

compareAll(Al, Bl; show, skip)

Recursively compare the all fields of T that are not in skip for objects Al and Bl.

TODO > add to func_ref.md

DistributedFactorGraphs.compareAllVariables — Method

compareAllVariables(fgA, fgB; skip, show, skipsamples)

Compare all variables in both ::FactorGraphs A and B.

Notes

A and B should all the same variables and factors.

Related:

compareFactorGraphs, compareSimilarVariables, compareVariable, ls

DistributedFactorGraphs.compareFactor — Method

compareFactor(A, B; show, skip, skipsamples, skipcompute)

Compare that all fields are the same in a ::FactorGraph factor.

DevNotes

TODO getSolverData(A).fnc.varValsAll / varidx are only defined downstream, so should should this function not be in IIF?

DistributedFactorGraphs.compareFactorGraphs — Method

compareFactorGraphs(
    fgA,
    fgB;
    skipsamples,
    skipcompute,
    skip,
    show
)

Compare and return if two factor graph objects are the same, by comparing similar variables and factors.

Notes:

Default items to skip with skipsamples, skipcompute.

User defined fields to skip can be specified with skip::Vector{Symbol}.

To enable debug messages for viewing which fields are not the same:

https://stackoverflow.com/questions/53548681/how-to-enable-debugging-messages-in-juno-julia-editor

Related:

compareSimilarVariables, compareSimilarFactors, compareAllVariables, ls.

DistributedFactorGraphs.compareFields — Method

compareFields(Al, Bl; show, skip)

Compare the all fields of T that are not in skip for objects Al and Bl and returns ::Bool.

TODO > add to func_ref.md

DistributedFactorGraphs.compareSimilarFactors — Method

compareSimilarFactors(
    fgA,
    fgB;
    skipsamples,
    skipcompute,
    skip,
    show
)

Compare similar factors between ::FactorGraphs A and B.

Related:

compareFactorGraphs, compareSimilarVariables, compareAllVariables, ls.

DistributedFactorGraphs.compareSimilarVariables — Method

compareSimilarVariables(fgA, fgB; skip, show, skipsamples)

Compare similar labels between ::FactorGraphs A and B.

Notes

At least one variable label should exist in both A and B.

Related:

compareFactorGraphs, compareAllVariables, compareSimilarFactors, compareVariable, ls.

DistributedFactorGraphs.compareVariable — Method

compareVariable(A, B; skip, show, skipsamples)

Compare that all fields are the same in a ::FactorGraph variable.

Common Functions

DistributedFactorGraphs.filterDFG! — Function

Filter nodes in a DFG based on a predicate function. This function modifies the input nodes vector in place, removing nodes that do not satisfy the predicate.

For cross-backend compatibility: Use only the standard predicates (==, <, <=, >, >=, in, contains, startswith, endswith) when you need your code to work with both in-memory and database-backed DFGs. These are likely to be supported by database query languages and are defined in std_numeric_predicates and std_string_predicates.

For in-memory only operations: You can use any Julia predicate, since you have full access to the data and Julia's capabilities. This is more flexible but will not work if you later switch to a database backend or another programming language.

Standard predicates

Numeric predicates: ==, <, <=, >, >=, in

String predicates: ==, contains, startswith, endswith, in

DistributedFactorGraphs.getVariableLabelNumber — Function

getVariableLabelNumber(vs)
getVariableLabelNumber(vs, prefix)

Small utility to return ::Int, e.g. 0 from getVariableLabelNumber(:x0)

Examples

getVariableLabelNumber(:l10)          # 10
getVariableLabelNumber(:x1)           # 1
getVariableLabelNumber(:x1_10, "x1_") # 10

DevNotes

make prefix Regex based for longer – i.e. :apriltag578, :lm1_4

DistributedFactorGraphs.isValidLabel — Method

isValidLabel(id)

Returns true if the label is valid for node.

DistributedFactorGraphs.solveGraph! — Function

Default non-parametric graph solution.

DistributedFactorGraphs.solveGraphParametric! — Function

Standard parametric graph solution (Experimental).

DistributedFactorGraphs.sortDFG — Method

sortDFG(vars; by, kwargs...)

Convenience wrapper for Base.sort. Sort variable (factor) lists in a meaningful way (by timestamp, label, etc), for example [:april;:x1_3;:x1_6;] Defaults to sorting by timestamp for variables and factors and using natural_lt for Symbols. See Base.sort for more detail.

Notes

Not fool proof, but does better than native sort.

Example

sortDFG(ls(dfg)) sortDFG(ls(dfg), by=getLabel, lt=natural_lt)

Related

ls, lsf

Serialization